Module pal::control_register::cr0

Control Register 0.

Contains system control flags that control operating mode and states of the processor.

Constants

LONG_NAME

NAME

Functions

am_is_disabled

Returns whether “Alignment Mask” is currently disabled.

am_is_disabled_in_value

Returns whether “Alignment Mask” is disabled in the given register value.

am_is_enabled

Returns whether “Alignment Mask” is currently enabled.

am_is_enabled_in_value

Returns whether “Alignment Mask” is enabled in the given register value.

cd_is_disabled

Returns whether “Cache Disable” is currently disabled.

cd_is_disabled_in_value

Returns whether “Cache Disable” is disabled in the given register value.

cd_is_enabled

Returns whether “Cache Disable” is currently enabled.

cd_is_enabled_in_value

Returns whether “Cache Disable” is enabled in the given register value.

disable_am

Disables “Alignment Mask”.

disable_am_in_value

Disables “Alignment Mask” in the given register value, returning the new value.

disable_cd

Disables “Cache Disable”.

disable_cd_in_value

Disables “Cache Disable” in the given register value, returning the new value.

disable_em

Disables “Emulation”.

disable_em_in_value

Disables “Emulation” in the given register value, returning the new value.

disable_et

Disables “Extension Type”.

disable_et_in_value

Disables “Extension Type” in the given register value, returning the new value.

disable_mp

Disables “Monitor Coprocessor”.

disable_mp_in_value

Disables “Monitor Coprocessor” in the given register value, returning the new value.

disable_ne

Disables “Numeric Error”.

disable_ne_in_value

Disables “Numeric Error” in the given register value, returning the new value.

disable_nw

Disables “Not Write-through”.

disable_nw_in_value

Disables “Not Write-through” in the given register value, returning the new value.

disable_pe

Disables “Protection Enable”.

disable_pe_in_value

Disables “Protection Enable” in the given register value, returning the new value.

disable_pg

Disables “Paging”.

disable_pg_in_value

Disables “Paging” in the given register value, returning the new value.

disable_ts

Disables “Task Switched”.

disable_ts_in_value

Disables “Task Switched” in the given register value, returning the new value.

disable_wp

Disables “Write Protect”.

disable_wp_in_value

Disables “Write Protect” in the given register value, returning the new value.

em_is_disabled

Returns whether “Emulation” is currently disabled.

em_is_disabled_in_value

Returns whether “Emulation” is disabled in the given register value.

em_is_enabled

Returns whether “Emulation” is currently enabled.

em_is_enabled_in_value

Returns whether “Emulation” is enabled in the given register value.

enable_am

Enables “Alignment Mask”.

enable_am_in_value

Enables “Alignment Mask” in the given register value, returning the new value.

enable_cd

Enables “Cache Disable”.

enable_cd_in_value

Enables “Cache Disable” in the given register value, returning the new value.

enable_em

Enables “Emulation”.

enable_em_in_value

Enables “Emulation” in the given register value, returning the new value.

enable_et

Enables “Extension Type”.

enable_et_in_value

Enables “Extension Type” in the given register value, returning the new value.

enable_mp

Enables “Monitor Coprocessor”.

enable_mp_in_value

Enables “Monitor Coprocessor” in the given register value, returning the new value.

enable_ne

Enables “Numeric Error”.

enable_ne_in_value

Enables “Numeric Error” in the given register value, returning the new value.

enable_nw

Enables “Not Write-through”.

enable_nw_in_value

Enables “Not Write-through” in the given register value, returning the new value.

enable_pe

Enables “Protection Enable”.

enable_pe_in_value

Enables “Protection Enable” in the given register value, returning the new value.

enable_pg

Enables “Paging”.

enable_pg_in_value

Enables “Paging” in the given register value, returning the new value.

enable_ts

Enables “Task Switched”.

enable_ts_in_value

Enables “Task Switched” in the given register value, returning the new value.

enable_wp

Enables “Write Protect”.

enable_wp_in_value

Enables “Write Protect” in the given register value, returning the new value.

et_is_disabled

Returns whether “Extension Type” is currently disabled.

et_is_disabled_in_value

Returns whether “Extension Type” is disabled in the given register value.

et_is_enabled

Returns whether “Extension Type” is currently enabled.

et_is_enabled_in_value

Returns whether “Extension Type” is enabled in the given register value.

get

Returns the current value of the register.

mp_is_disabled

Returns whether “Monitor Coprocessor” is currently disabled.

mp_is_disabled_in_value

Returns whether “Monitor Coprocessor” is disabled in the given register value.

mp_is_enabled

Returns whether “Monitor Coprocessor” is currently enabled.

mp_is_enabled_in_value

Returns whether “Monitor Coprocessor” is enabled in the given register value.

ne_is_disabled

Returns whether “Numeric Error” is currently disabled.

ne_is_disabled_in_value

Returns whether “Numeric Error” is disabled in the given register value.

ne_is_enabled

Returns whether “Numeric Error” is currently enabled.

ne_is_enabled_in_value

Returns whether “Numeric Error” is enabled in the given register value.

nw_is_disabled

Returns whether “Not Write-through” is currently disabled.

nw_is_disabled_in_value

Returns whether “Not Write-through” is disabled in the given register value.

nw_is_enabled

Returns whether “Not Write-through” is currently enabled.

nw_is_enabled_in_value

Returns whether “Not Write-through” is enabled in the given register value.

pe_is_disabled

Returns whether “Protection Enable” is currently disabled.

pe_is_disabled_in_value

Returns whether “Protection Enable” is disabled in the given register value.

pe_is_enabled

Returns whether “Protection Enable” is currently enabled.

pe_is_enabled_in_value

Returns whether “Protection Enable” is enabled in the given register value.

pg_is_disabled

Returns whether “Paging” is currently disabled.

pg_is_disabled_in_value

Returns whether “Paging” is disabled in the given register value.

pg_is_enabled

Returns whether “Paging” is currently enabled.

pg_is_enabled_in_value

Returns whether “Paging” is enabled in the given register value.

set

Sets the value of the register.

ts_is_disabled

Returns whether “Task Switched” is currently disabled.

ts_is_disabled_in_value

Returns whether “Task Switched” is disabled in the given register value.

ts_is_enabled

Returns whether “Task Switched” is currently enabled.

ts_is_enabled_in_value

Returns whether “Task Switched” is enabled in the given register value.

wp_is_disabled

Returns whether “Write Protect” is currently disabled.

wp_is_disabled_in_value

Returns whether “Write Protect” is disabled in the given register value.

wp_is_enabled

Returns whether “Write Protect” is currently enabled.

wp_is_enabled_in_value

Returns whether “Write Protect” is enabled in the given register value.