Module pal::control_register::cr3

Control Register 3.

Contains the physical address of the base of the paging-structure hierarchy and two flags (PCD and PWT). Only the most-significant bits (less the lower 12 bits) of the base address are specified; the lower 12 bits of the address are assumed to be 0. The first paging structure must thus be aligned to a page (4-KByte) boundary. The PCD and PWT flags control caching of that paging structure in the processor’s internal data caches (they do not control TLB caching of page-directory information)

Constants

LONG_NAME

NAME

Functions

disable_pcd

Disables “Page-level Cache Disable”.

disable_pcd_in_value

Disables “Page-level Cache Disable” in the given register value, returning the new value.

disable_pwt

Disables “Page-level Write-Through”.

disable_pwt_in_value

Disables “Page-level Write-Through” in the given register value, returning the new value.

enable_pcd

Enables “Page-level Cache Disable”.

enable_pcd_in_value

Enables “Page-level Cache Disable” in the given register value, returning the new value.

enable_pwt

Enables “Page-level Write-Through”.

enable_pwt_in_value

Enables “Page-level Write-Through” in the given register value, returning the new value.

get

Returns the current value of the register.

get_page_directory_base

Returns the current value of “Page-Directory Base”.

get_page_directory_base_from_value

Returns the value of “Page-Directory Base” in the given register value.

pcd_is_disabled

Returns whether “Page-level Cache Disable” is currently disabled.

pcd_is_disabled_in_value

Returns whether “Page-level Cache Disable” is disabled in the given register value.

pcd_is_enabled

Returns whether “Page-level Cache Disable” is currently enabled.

pcd_is_enabled_in_value

Returns whether “Page-level Cache Disable” is enabled in the given register value.

pwt_is_disabled

Returns whether “Page-level Write-Through” is currently disabled.

pwt_is_disabled_in_value

Returns whether “Page-level Write-Through” is disabled in the given register value.

pwt_is_enabled

Returns whether “Page-level Write-Through” is currently enabled.

pwt_is_enabled_in_value

Returns whether “Page-level Write-Through” is enabled in the given register value.

set

Sets the value of the register.

set_page_directory_base

Sets the value of “Page-Directory Base”.

set_page_directory_base_in_value

Sets the value of “Page-Directory Base” in the given register value, returning the new value.