Module pal::control_register::cr4

Control Register 4.

Contains a group of flags that enable several architectural extensions, and indicate operating system or executive support for specific processor capabilities

Constants

LONG_NAME

NAME

Functions

de_is_disabled

Returns whether “Debugging Extensions” is currently disabled.

de_is_disabled_in_value

Returns whether “Debugging Extensions” is disabled in the given register value.

de_is_enabled

Returns whether “Debugging Extensions” is currently enabled.

de_is_enabled_in_value

Returns whether “Debugging Extensions” is enabled in the given register value.

disable_de

Disables “Debugging Extensions”.

disable_de_in_value

Disables “Debugging Extensions” in the given register value, returning the new value.

disable_fsgsbase

Disables “FSGSBASE-Enable Bit”.

disable_fsgsbase_in_value

Disables “FSGSBASE-Enable Bit” in the given register value, returning the new value.

disable_mce

Disables “Machine-Check Enable”.

disable_mce_in_value

Disables “Machine-Check Enable” in the given register value, returning the new value.

disable_osfxsr

Disables “Operating System Support for FXSAVE and FXRSTOR instructions”.

disable_osfxsr_in_value

Disables “Operating System Support for FXSAVE and FXRSTOR instructions” in the given register value, returning the new value.

disable_osxmmexcpt

Disables “Operating System Support for Unmasked SIMD Floating-PointExceptions”.

disable_osxmmexcpt_in_value

Disables “Operating System Support for Unmasked SIMD Floating-PointExceptions” in the given register value, returning the new value.

disable_osxsave

Disables “XSAVE and Processor Extended States-Enable Bit”.

disable_osxsave_in_value

Disables “XSAVE and Processor Extended States-Enable Bit” in the given register value, returning the new value.

disable_pae

Disables “Physical Address Extension”.

disable_pae_in_value

Disables “Physical Address Extension” in the given register value, returning the new value.

disable_pce

Disables “Performance-Monitoring Counter Enable”.

disable_pce_in_value

Disables “Performance-Monitoring Counter Enable” in the given register value, returning the new value.

disable_pcide

Disables “PCID-Enable Bit”.

disable_pcide_in_value

Disables “PCID-Enable Bit” in the given register value, returning the new value.

disable_pge

Disables “Page Global Enable”.

disable_pge_in_value

Disables “Page Global Enable” in the given register value, returning the new value.

disable_pke

Disables “Protection-Key-Enable Bit”.

disable_pke_in_value

Disables “Protection-Key-Enable Bit” in the given register value, returning the new value.

disable_pse

Disables “Page Size Extensions”.

disable_pse_in_value

Disables “Page Size Extensions” in the given register value, returning the new value.

disable_pvi

Disables “Protected-Mode Virtual Interrupts”.

disable_pvi_in_value

Disables “Protected-Mode Virtual Interrupts” in the given register value, returning the new value.

disable_smap

Disables “SMAP-Enable Bit”.

disable_smap_in_value

Disables “SMAP-Enable Bit” in the given register value, returning the new value.

disable_smep

Disables “SMEP-Enable Bit”.

disable_smep_in_value

Disables “SMEP-Enable Bit” in the given register value, returning the new value.

disable_smxe

Disables “SMX-Enable Bit”.

disable_smxe_in_value

Disables “SMX-Enable Bit” in the given register value, returning the new value.

disable_tsd

Disables “Time Stamp Disable”.

disable_tsd_in_value

Disables “Time Stamp Disable” in the given register value, returning the new value.

disable_umip

Disables “User-Mode Instruction Prevention”.

disable_umip_in_value

Disables “User-Mode Instruction Prevention” in the given register value, returning the new value.

disable_vme

Disables “Virtual-8086 Mode Extensions”.

disable_vme_in_value

Disables “Virtual-8086 Mode Extensions” in the given register value, returning the new value.

disable_vmxe

Disables “VMX-Enable Bit”.

disable_vmxe_in_value

Disables “VMX-Enable Bit” in the given register value, returning the new value.

enable_de

Enables “Debugging Extensions”.

enable_de_in_value

Enables “Debugging Extensions” in the given register value, returning the new value.

enable_fsgsbase

Enables “FSGSBASE-Enable Bit”.

enable_fsgsbase_in_value

Enables “FSGSBASE-Enable Bit” in the given register value, returning the new value.

enable_mce

Enables “Machine-Check Enable”.

enable_mce_in_value

Enables “Machine-Check Enable” in the given register value, returning the new value.

enable_osfxsr

Enables “Operating System Support for FXSAVE and FXRSTOR instructions”.

enable_osfxsr_in_value

Enables “Operating System Support for FXSAVE and FXRSTOR instructions” in the given register value, returning the new value.

enable_osxmmexcpt

Enables “Operating System Support for Unmasked SIMD Floating-PointExceptions”.

enable_osxmmexcpt_in_value

Enables “Operating System Support for Unmasked SIMD Floating-PointExceptions” in the given register value, returning the new value.

enable_osxsave

Enables “XSAVE and Processor Extended States-Enable Bit”.

enable_osxsave_in_value

Enables “XSAVE and Processor Extended States-Enable Bit” in the given register value, returning the new value.

enable_pae

Enables “Physical Address Extension”.

enable_pae_in_value

Enables “Physical Address Extension” in the given register value, returning the new value.

enable_pce

Enables “Performance-Monitoring Counter Enable”.

enable_pce_in_value

Enables “Performance-Monitoring Counter Enable” in the given register value, returning the new value.

enable_pcide

Enables “PCID-Enable Bit”.

enable_pcide_in_value

Enables “PCID-Enable Bit” in the given register value, returning the new value.

enable_pge

Enables “Page Global Enable”.

enable_pge_in_value

Enables “Page Global Enable” in the given register value, returning the new value.

enable_pke

Enables “Protection-Key-Enable Bit”.

enable_pke_in_value

Enables “Protection-Key-Enable Bit” in the given register value, returning the new value.

enable_pse

Enables “Page Size Extensions”.

enable_pse_in_value

Enables “Page Size Extensions” in the given register value, returning the new value.

enable_pvi

Enables “Protected-Mode Virtual Interrupts”.

enable_pvi_in_value

Enables “Protected-Mode Virtual Interrupts” in the given register value, returning the new value.

enable_smap

Enables “SMAP-Enable Bit”.

enable_smap_in_value

Enables “SMAP-Enable Bit” in the given register value, returning the new value.

enable_smep

Enables “SMEP-Enable Bit”.

enable_smep_in_value

Enables “SMEP-Enable Bit” in the given register value, returning the new value.

enable_smxe

Enables “SMX-Enable Bit”.

enable_smxe_in_value

Enables “SMX-Enable Bit” in the given register value, returning the new value.

enable_tsd

Enables “Time Stamp Disable”.

enable_tsd_in_value

Enables “Time Stamp Disable” in the given register value, returning the new value.

enable_umip

Enables “User-Mode Instruction Prevention”.

enable_umip_in_value

Enables “User-Mode Instruction Prevention” in the given register value, returning the new value.

enable_vme

Enables “Virtual-8086 Mode Extensions”.

enable_vme_in_value

Enables “Virtual-8086 Mode Extensions” in the given register value, returning the new value.

enable_vmxe

Enables “VMX-Enable Bit”.

enable_vmxe_in_value

Enables “VMX-Enable Bit” in the given register value, returning the new value.

fsgsbase_is_disabled

Returns whether “FSGSBASE-Enable Bit” is currently disabled.

fsgsbase_is_disabled_in_value

Returns whether “FSGSBASE-Enable Bit” is disabled in the given register value.

fsgsbase_is_enabled

Returns whether “FSGSBASE-Enable Bit” is currently enabled.

fsgsbase_is_enabled_in_value

Returns whether “FSGSBASE-Enable Bit” is enabled in the given register value.

get

Returns the current value of the register.

mce_is_disabled

Returns whether “Machine-Check Enable” is currently disabled.

mce_is_disabled_in_value

Returns whether “Machine-Check Enable” is disabled in the given register value.

mce_is_enabled

Returns whether “Machine-Check Enable” is currently enabled.

mce_is_enabled_in_value

Returns whether “Machine-Check Enable” is enabled in the given register value.

osfxsr_is_disabled

Returns whether “Operating System Support for FXSAVE and FXRSTOR instructions” is currently disabled.

osfxsr_is_disabled_in_value

Returns whether “Operating System Support for FXSAVE and FXRSTOR instructions” is disabled in the given register value.

osfxsr_is_enabled

Returns whether “Operating System Support for FXSAVE and FXRSTOR instructions” is currently enabled.

osfxsr_is_enabled_in_value

Returns whether “Operating System Support for FXSAVE and FXRSTOR instructions” is enabled in the given register value.

osxmmexcpt_is_disabled

Returns whether “Operating System Support for Unmasked SIMD Floating-PointExceptions” is currently disabled.

osxmmexcpt_is_disabled_in_value

Returns whether “Operating System Support for Unmasked SIMD Floating-PointExceptions” is disabled in the given register value.

osxmmexcpt_is_enabled

Returns whether “Operating System Support for Unmasked SIMD Floating-PointExceptions” is currently enabled.

osxmmexcpt_is_enabled_in_value

Returns whether “Operating System Support for Unmasked SIMD Floating-PointExceptions” is enabled in the given register value.

osxsave_is_disabled

Returns whether “XSAVE and Processor Extended States-Enable Bit” is currently disabled.

osxsave_is_disabled_in_value

Returns whether “XSAVE and Processor Extended States-Enable Bit” is disabled in the given register value.

osxsave_is_enabled

Returns whether “XSAVE and Processor Extended States-Enable Bit” is currently enabled.

osxsave_is_enabled_in_value

Returns whether “XSAVE and Processor Extended States-Enable Bit” is enabled in the given register value.

pae_is_disabled

Returns whether “Physical Address Extension” is currently disabled.

pae_is_disabled_in_value

Returns whether “Physical Address Extension” is disabled in the given register value.

pae_is_enabled

Returns whether “Physical Address Extension” is currently enabled.

pae_is_enabled_in_value

Returns whether “Physical Address Extension” is enabled in the given register value.

pce_is_disabled

Returns whether “Performance-Monitoring Counter Enable” is currently disabled.

pce_is_disabled_in_value

Returns whether “Performance-Monitoring Counter Enable” is disabled in the given register value.

pce_is_enabled

Returns whether “Performance-Monitoring Counter Enable” is currently enabled.

pce_is_enabled_in_value

Returns whether “Performance-Monitoring Counter Enable” is enabled in the given register value.

pcide_is_disabled

Returns whether “PCID-Enable Bit” is currently disabled.

pcide_is_disabled_in_value

Returns whether “PCID-Enable Bit” is disabled in the given register value.

pcide_is_enabled

Returns whether “PCID-Enable Bit” is currently enabled.

pcide_is_enabled_in_value

Returns whether “PCID-Enable Bit” is enabled in the given register value.

pge_is_disabled

Returns whether “Page Global Enable” is currently disabled.

pge_is_disabled_in_value

Returns whether “Page Global Enable” is disabled in the given register value.

pge_is_enabled

Returns whether “Page Global Enable” is currently enabled.

pge_is_enabled_in_value

Returns whether “Page Global Enable” is enabled in the given register value.

pke_is_disabled

Returns whether “Protection-Key-Enable Bit” is currently disabled.

pke_is_disabled_in_value

Returns whether “Protection-Key-Enable Bit” is disabled in the given register value.

pke_is_enabled

Returns whether “Protection-Key-Enable Bit” is currently enabled.

pke_is_enabled_in_value

Returns whether “Protection-Key-Enable Bit” is enabled in the given register value.

pse_is_disabled

Returns whether “Page Size Extensions” is currently disabled.

pse_is_disabled_in_value

Returns whether “Page Size Extensions” is disabled in the given register value.

pse_is_enabled

Returns whether “Page Size Extensions” is currently enabled.

pse_is_enabled_in_value

Returns whether “Page Size Extensions” is enabled in the given register value.

pvi_is_disabled

Returns whether “Protected-Mode Virtual Interrupts” is currently disabled.

pvi_is_disabled_in_value

Returns whether “Protected-Mode Virtual Interrupts” is disabled in the given register value.

pvi_is_enabled

Returns whether “Protected-Mode Virtual Interrupts” is currently enabled.

pvi_is_enabled_in_value

Returns whether “Protected-Mode Virtual Interrupts” is enabled in the given register value.

set

Sets the value of the register.

smap_is_disabled

Returns whether “SMAP-Enable Bit” is currently disabled.

smap_is_disabled_in_value

Returns whether “SMAP-Enable Bit” is disabled in the given register value.

smap_is_enabled

Returns whether “SMAP-Enable Bit” is currently enabled.

smap_is_enabled_in_value

Returns whether “SMAP-Enable Bit” is enabled in the given register value.

smep_is_disabled

Returns whether “SMEP-Enable Bit” is currently disabled.

smep_is_disabled_in_value

Returns whether “SMEP-Enable Bit” is disabled in the given register value.

smep_is_enabled

Returns whether “SMEP-Enable Bit” is currently enabled.

smep_is_enabled_in_value

Returns whether “SMEP-Enable Bit” is enabled in the given register value.

smxe_is_disabled

Returns whether “SMX-Enable Bit” is currently disabled.

smxe_is_disabled_in_value

Returns whether “SMX-Enable Bit” is disabled in the given register value.

smxe_is_enabled

Returns whether “SMX-Enable Bit” is currently enabled.

smxe_is_enabled_in_value

Returns whether “SMX-Enable Bit” is enabled in the given register value.

tsd_is_disabled

Returns whether “Time Stamp Disable” is currently disabled.

tsd_is_disabled_in_value

Returns whether “Time Stamp Disable” is disabled in the given register value.

tsd_is_enabled

Returns whether “Time Stamp Disable” is currently enabled.

tsd_is_enabled_in_value

Returns whether “Time Stamp Disable” is enabled in the given register value.

umip_is_disabled

Returns whether “User-Mode Instruction Prevention” is currently disabled.

umip_is_disabled_in_value

Returns whether “User-Mode Instruction Prevention” is disabled in the given register value.

umip_is_enabled

Returns whether “User-Mode Instruction Prevention” is currently enabled.

umip_is_enabled_in_value

Returns whether “User-Mode Instruction Prevention” is enabled in the given register value.

vme_is_disabled

Returns whether “Virtual-8086 Mode Extensions” is currently disabled.

vme_is_disabled_in_value

Returns whether “Virtual-8086 Mode Extensions” is disabled in the given register value.

vme_is_enabled

Returns whether “Virtual-8086 Mode Extensions” is currently enabled.

vme_is_enabled_in_value

Returns whether “Virtual-8086 Mode Extensions” is enabled in the given register value.

vmxe_is_disabled

Returns whether “VMX-Enable Bit” is currently disabled.

vmxe_is_disabled_in_value

Returns whether “VMX-Enable Bit” is disabled in the given register value.

vmxe_is_enabled

Returns whether “VMX-Enable Bit” is currently enabled.

vmxe_is_enabled_in_value

Returns whether “VMX-Enable Bit” is enabled in the given register value.