Module pal::control_register::xcr0

Extended Control Register 0.

If CPUID.01H:ECX.XSAVE[bit 26] is 1, the processor supports one or more extended control registers (XCRs). Currently, the only such register defined is XCR0. This register specifies the set of processor state components for which the operating system provides context management, e.g. x87 FPU state, SSE state, AVX state. The OS programs XCR0 to reflect the features for which it provides context management.

Constants

LONG_NAME

NAME

Functions

avx_is_disabled

Returns whether “XCR0.AVX (bit 2)” is currently disabled.

avx_is_disabled_in_value

Returns whether “XCR0.AVX (bit 2)” is disabled in the given register value.

avx_is_enabled

Returns whether “XCR0.AVX (bit 2)” is currently enabled.

avx_is_enabled_in_value

Returns whether “XCR0.AVX (bit 2)” is enabled in the given register value.

bndcsr_is_disabled

Returns whether “XCR0.BNDCSR (bit 4)” is currently disabled.

bndcsr_is_disabled_in_value

Returns whether “XCR0.BNDCSR (bit 4)” is disabled in the given register value.

bndcsr_is_enabled

Returns whether “XCR0.BNDCSR (bit 4)” is currently enabled.

bndcsr_is_enabled_in_value

Returns whether “XCR0.BNDCSR (bit 4)” is enabled in the given register value.

bndreg_is_disabled

Returns whether “XCR0.BNDREG (bit 3)” is currently disabled.

bndreg_is_disabled_in_value

Returns whether “XCR0.BNDREG (bit 3)” is disabled in the given register value.

bndreg_is_enabled

Returns whether “XCR0.BNDREG (bit 3)” is currently enabled.

bndreg_is_enabled_in_value

Returns whether “XCR0.BNDREG (bit 3)” is enabled in the given register value.

disable_avx

Disables “XCR0.AVX (bit 2)”.

disable_avx_in_value

Disables “XCR0.AVX (bit 2)” in the given register value, returning the new value.

disable_bndcsr

Disables “XCR0.BNDCSR (bit 4)”.

disable_bndcsr_in_value

Disables “XCR0.BNDCSR (bit 4)” in the given register value, returning the new value.

disable_bndreg

Disables “XCR0.BNDREG (bit 3)”.

disable_bndreg_in_value

Disables “XCR0.BNDREG (bit 3)” in the given register value, returning the new value.

disable_hi16_zmm

Disables “XCR0.Hi16_ZMM (bit 7)”.

disable_hi16_zmm_in_value

Disables “XCR0.Hi16_ZMM (bit 7)” in the given register value, returning the new value.

disable_opmask

Disables “XCR0.opmask (bit 5)”.

disable_opmask_in_value

Disables “XCR0.opmask (bit 5)” in the given register value, returning the new value.

disable_pkru

Disables “XCR0.PKRU (bit 9)”.

disable_pkru_in_value

Disables “XCR0.PKRU (bit 9)” in the given register value, returning the new value.

disable_sse

Disables “XCR0.SSE (bit 1)”.

disable_sse_in_value

Disables “XCR0.SSE (bit 1)” in the given register value, returning the new value.

disable_x87

Disables “XCR0.X87 (bit 0)”.

disable_x87_in_value

Disables “XCR0.X87 (bit 0)” in the given register value, returning the new value.

disable_zmm_hi256

Disables “XCR0.ZMM_Hi256 (bit 6)”.

disable_zmm_hi256_in_value

Disables “XCR0.ZMM_Hi256 (bit 6)” in the given register value, returning the new value.

enable_avx

Enables “XCR0.AVX (bit 2)”.

enable_avx_in_value

Enables “XCR0.AVX (bit 2)” in the given register value, returning the new value.

enable_bndcsr

Enables “XCR0.BNDCSR (bit 4)”.

enable_bndcsr_in_value

Enables “XCR0.BNDCSR (bit 4)” in the given register value, returning the new value.

enable_bndreg

Enables “XCR0.BNDREG (bit 3)”.

enable_bndreg_in_value

Enables “XCR0.BNDREG (bit 3)” in the given register value, returning the new value.

enable_hi16_zmm

Enables “XCR0.Hi16_ZMM (bit 7)”.

enable_hi16_zmm_in_value

Enables “XCR0.Hi16_ZMM (bit 7)” in the given register value, returning the new value.

enable_opmask

Enables “XCR0.opmask (bit 5)”.

enable_opmask_in_value

Enables “XCR0.opmask (bit 5)” in the given register value, returning the new value.

enable_pkru

Enables “XCR0.PKRU (bit 9)”.

enable_pkru_in_value

Enables “XCR0.PKRU (bit 9)” in the given register value, returning the new value.

enable_sse

Enables “XCR0.SSE (bit 1)”.

enable_sse_in_value

Enables “XCR0.SSE (bit 1)” in the given register value, returning the new value.

enable_x87

Enables “XCR0.X87 (bit 0)”.

enable_x87_in_value

Enables “XCR0.X87 (bit 0)” in the given register value, returning the new value.

enable_zmm_hi256

Enables “XCR0.ZMM_Hi256 (bit 6)”.

enable_zmm_hi256_in_value

Enables “XCR0.ZMM_Hi256 (bit 6)” in the given register value, returning the new value.

get

Returns the current value of the register.

hi16_zmm_is_disabled

Returns whether “XCR0.Hi16_ZMM (bit 7)” is currently disabled.

hi16_zmm_is_disabled_in_value

Returns whether “XCR0.Hi16_ZMM (bit 7)” is disabled in the given register value.

hi16_zmm_is_enabled

Returns whether “XCR0.Hi16_ZMM (bit 7)” is currently enabled.

hi16_zmm_is_enabled_in_value

Returns whether “XCR0.Hi16_ZMM (bit 7)” is enabled in the given register value.

opmask_is_disabled

Returns whether “XCR0.opmask (bit 5)” is currently disabled.

opmask_is_disabled_in_value

Returns whether “XCR0.opmask (bit 5)” is disabled in the given register value.

opmask_is_enabled

Returns whether “XCR0.opmask (bit 5)” is currently enabled.

opmask_is_enabled_in_value

Returns whether “XCR0.opmask (bit 5)” is enabled in the given register value.

pkru_is_disabled

Returns whether “XCR0.PKRU (bit 9)” is currently disabled.

pkru_is_disabled_in_value

Returns whether “XCR0.PKRU (bit 9)” is disabled in the given register value.

pkru_is_enabled

Returns whether “XCR0.PKRU (bit 9)” is currently enabled.

pkru_is_enabled_in_value

Returns whether “XCR0.PKRU (bit 9)” is enabled in the given register value.

set

Sets the value of the register.

sse_is_disabled

Returns whether “XCR0.SSE (bit 1)” is currently disabled.

sse_is_disabled_in_value

Returns whether “XCR0.SSE (bit 1)” is disabled in the given register value.

sse_is_enabled

Returns whether “XCR0.SSE (bit 1)” is currently enabled.

sse_is_enabled_in_value

Returns whether “XCR0.SSE (bit 1)” is enabled in the given register value.

x87_is_disabled

Returns whether “XCR0.X87 (bit 0)” is currently disabled.

x87_is_disabled_in_value

Returns whether “XCR0.X87 (bit 0)” is disabled in the given register value.

x87_is_enabled

Returns whether “XCR0.X87 (bit 0)” is currently enabled.

x87_is_enabled_in_value

Returns whether “XCR0.X87 (bit 0)” is enabled in the given register value.

zmm_hi256_is_disabled

Returns whether “XCR0.ZMM_Hi256 (bit 6)” is currently disabled.

zmm_hi256_is_disabled_in_value

Returns whether “XCR0.ZMM_Hi256 (bit 6)” is disabled in the given register value.

zmm_hi256_is_enabled

Returns whether “XCR0.ZMM_Hi256 (bit 6)” is currently enabled.

zmm_hi256_is_enabled_in_value

Returns whether “XCR0.ZMM_Hi256 (bit 6)” is enabled in the given register value.