Module pal::cpuid

Re-exports

pub use leaf_01_eax::*;

pub use leaf_01_ebx::*;

pub use leaf_01_ecx::*;

pub use leaf_01_edx::*;

pub use leaf_02_eax::*;

pub use leaf_02_ebx::*;

pub use leaf_02_ecx::*;

pub use leaf_02_edx::*;

pub use leaf_03_eax::*;

pub use leaf_03_ebx::*;

pub use leaf_03_ecx::*;

pub use leaf_03_edx::*;

pub use leaf_04_eax::*;

pub use leaf_04_ebx::*;

pub use leaf_04_ecx::*;

pub use leaf_04_edx::*;

pub use leaf_05_eax::*;

pub use leaf_05_ebx::*;

pub use leaf_05_ecx::*;

pub use leaf_05_edx::*;

pub use leaf_06_eax::*;

pub use leaf_06_ebx::*;

pub use leaf_06_ecx::*;

pub use leaf_06_edx::*;

pub use leaf_07_eax::*;

pub use leaf_07_ebx::*;

pub use leaf_07_ecx::*;

pub use leaf_07_edx::*;

pub use leaf_09_eax::*;

pub use leaf_09_ebx::*;

pub use leaf_09_ecx::*;

pub use leaf_09_edx::*;

pub use leaf_0a_eax::*;

pub use leaf_0a_ebx::*;

pub use leaf_0a_ecx::*;

pub use leaf_0a_edx::*;

pub use leaf_0b_eax::*;

pub use leaf_0b_ebx::*;

pub use leaf_0b_ecx::*;

pub use leaf_0b_edx::*;

pub use leaf_0d_eax::*;

pub use leaf_0d_ebx::*;

pub use leaf_0d_ecx::*;

pub use leaf_0d_edx::*;

pub use leaf_0f_eax::*;

pub use leaf_0f_ebx::*;

pub use leaf_0f_ecx::*;

pub use leaf_0f_edx::*;

pub use leaf_10_eax::*;

pub use leaf_10_ebx::*;

pub use leaf_10_ecx::*;

pub use leaf_10_edx::*;

pub use leaf_12_eax::*;

pub use leaf_12_ebx::*;

pub use leaf_12_ecx::*;

pub use leaf_12_edx::*;

pub use leaf_14_eax::*;

pub use leaf_14_ebx::*;

pub use leaf_14_ecx::*;

pub use leaf_14_edx::*;

pub use leaf_15_eax::*;

pub use leaf_15_ebx::*;

pub use leaf_15_ecx::*;

pub use leaf_15_edx::*;

pub use leaf_16_eax::*;

pub use leaf_16_ebx::*;

pub use leaf_16_ecx::*;

pub use leaf_16_edx::*;

pub use leaf_17_eax::*;

pub use leaf_17_ebx::*;

pub use leaf_17_ecx::*;

pub use leaf_17_edx::*;

pub use leaf_18_eax::*;

pub use leaf_18_ebx::*;

pub use leaf_18_ecx::*;

pub use leaf_18_edx::*;

pub use leaf_1a_eax::*;

pub use leaf_1a_ebx::*;

pub use leaf_1a_ecx::*;

pub use leaf_1a_edx::*;

pub use leaf_1f_eax::*;

pub use leaf_1f_ebx::*;

pub use leaf_1f_ecx::*;

pub use leaf_1f_edx::*;

pub use leaf_80000000_eax::*;

pub use leaf_80000000_ebx::*;

pub use leaf_80000000_ecx::*;

pub use leaf_80000000_edx::*;

pub use leaf_80000001_eax::*;

pub use leaf_80000001_ebx::*;

pub use leaf_80000001_ecx::*;

pub use leaf_80000001_edx::*;

pub use leaf_80000002_eax::*;

pub use leaf_80000002_ebx::*;

pub use leaf_80000002_ecx::*;

pub use leaf_80000002_edx::*;

pub use leaf_80000003_eax::*;

pub use leaf_80000003_ebx::*;

pub use leaf_80000003_ecx::*;

pub use leaf_80000003_edx::*;

pub use leaf_80000004_eax::*;

pub use leaf_80000004_ebx::*;

pub use leaf_80000004_ecx::*;

pub use leaf_80000004_edx::*;

pub use leaf_80000006_eax::*;

pub use leaf_80000006_ebx::*;

pub use leaf_80000006_ecx::*;

pub use leaf_80000006_edx::*;

pub use leaf_80000007_eax::*;

pub use leaf_80000007_ebx::*;

pub use leaf_80000007_ecx::*;

pub use leaf_80000007_edx::*;

pub use leaf_80000008_eax::*;

pub use leaf_80000008_ebx::*;

pub use leaf_80000008_ecx::*;

pub use leaf_80000008_edx::*;

Modules

leaf_0a_eax

Architectural Performance Monitoring Leaf.

leaf_0a_ebx

Architectural Performance Monitoring Leaf.

leaf_0a_ecx

Architectural Performance Monitoring Leaf.

leaf_0a_edx

Architectural Performance Monitoring Leaf.

leaf_0b_eax

Extended Topology Enumeration Leaf.

leaf_0b_ebx

Extended Topology Enumeration Leaf.

leaf_0b_ecx

Extended Topology Enumeration Leaf.

leaf_0b_edx

Extended Topology Enumeration Leaf.

leaf_0d_eax

Processor Extended State Enumeration.

leaf_0d_ebx

Processor Extended State Enumeration.

leaf_0d_ecx

Processor Extended State Enumeration.

leaf_0d_edx

Processor Extended State Enumeration.

leaf_0f_eax

Returns Intel Resource Director Technology (Intel RDT) MonitoringEnumeration Information.

leaf_0f_ebx

Returns Intel Resource Director Technology (Intel RDT) MonitoringEnumeration Information.

leaf_0f_ecx

Returns Intel Resource Director Technology (Intel RDT) MonitoringEnumeration Information.

leaf_0f_edx

Returns Intel Resource Director Technology (Intel RDT) MonitoringEnumeration Information.

leaf_01_eax

Basic CPUID Information.

leaf_01_ebx

Basic CPUID Information.

leaf_01_ecx

Basic CPUID Information.

leaf_01_edx

Basic CPUID Information.

leaf_1a_eax

Returns Hybrid Information.

leaf_1a_ebx

Returns Hybrid Information.

leaf_1a_ecx

Returns Hybrid Information.

leaf_1a_edx

Returns Hybrid Information.

leaf_1f_eax

Returns V2 Extended Topology Information.

leaf_1f_ebx

Returns V2 Extended Topology Information.

leaf_1f_ecx

Returns V2 Extended Topology Information.

leaf_1f_edx

Returns V2 Extended Topology Information.

leaf_02_eax

Basic CPUID Information.

leaf_02_ebx

Basic CPUID Information.

leaf_02_ecx

Basic CPUID Information.

leaf_02_edx

Basic CPUID Information.

leaf_03_eax

Basic CPUID Information.

leaf_03_ebx

Basic CPUID Information.

leaf_03_ecx

Basic CPUID Information.

leaf_03_edx

Basic CPUID Information.

leaf_04_eax

Deterministic Cache Parameters Leaf.

leaf_04_ebx

Deterministic Cache Parameters Leaf.

leaf_04_ecx

Deterministic Cache Parameters Leaf.

leaf_04_edx

Deterministic Cache Parameters Leaf.

leaf_05_eax

MONITOR/MWAIT Leaf.

leaf_05_ebx

MONITOR/MWAIT Leaf.

leaf_05_ecx

MONITOR/MWAIT Leaf.

leaf_05_edx

MONITOR/MWAIT Leaf.

leaf_06_eax

Thermal and Power Management Leaf.

leaf_06_ebx

Thermal and Power Management Leaf.

leaf_06_ecx

Thermal and Power Management Leaf.

leaf_06_edx

Thermal and Power Management Leaf.

leaf_07_eax

Structured Extended Feature Flags Enumeration Leaf.

leaf_07_ebx

Structured Extended Feature Flags Enumeration Leaf.

leaf_07_ecx

Structured Extended Feature Flags Enumeration Leaf.

leaf_07_edx

Structured Extended Feature Flags Enumeration Leaf.

leaf_09_eax

Direct Cache Access Information Leaf.

leaf_09_ebx

Direct Cache Access Information Leaf.

leaf_09_ecx

Direct Cache Access Information Leaf.

leaf_09_edx

Direct Cache Access Information Leaf.

leaf_10_eax

Returns Intel Resource Director Technology (Intel RDT) AllocationEnumeration Information.

leaf_10_ebx

Returns Intel Resource Director Technology (Intel RDT) AllocationEnumeration Information.

leaf_10_ecx

Returns Intel Resource Director Technology (Intel RDT) AllocationEnumeration Information.

leaf_10_edx

Returns Intel Resource Director Technology (Intel RDT) AllocationEnumeration Information.

leaf_12_eax

Returns Intel SGX Enumeration Information.

leaf_12_ebx

Returns Intel SGX Enumeration Information.

leaf_12_ecx

Returns Intel SGX Enumeration Information.

leaf_12_edx

Returns Intel SGX Enumeration Information.

leaf_14_eax

Returns Intel Processor Trace Enumeration Information.

leaf_14_ebx

Returns Intel Processor Trace Enumeration Information.

leaf_14_ecx

Returns Intel Processor Trace Enumeration Information.

leaf_14_edx

Returns Intel Processor Trace Enumeration Information.

leaf_15_eax

Returns Time Stamp Counter and Nominal Core Crystal Clock Information.

leaf_15_ebx

Returns Time Stamp Counter and Nominal Core Crystal Clock Information.

leaf_15_ecx

Returns Time Stamp Counter and Nominal Core Crystal Clock Information.

leaf_15_edx

Returns Time Stamp Counter and Nominal Core Crystal Clock Information.

leaf_16_eax

Returns Processor Frequency Information.

leaf_16_ebx

Returns Processor Frequency Information.

leaf_16_ecx

Returns Processor Frequency Information.

leaf_16_edx

Returns Processor Frequency Information.

leaf_17_eax

Returns System-On-Chip Information.

leaf_17_ebx

Returns System-On-Chip Information.

leaf_17_ecx

Returns System-On-Chip Information.

leaf_17_edx

Returns System-On-Chip Information.

leaf_18_eax

Returns Deterministic Address Translation Parameters Information.

leaf_18_ebx

Returns Deterministic Address Translation Parameters Information.

leaf_18_ecx

Returns Deterministic Address Translation Parameters Information.

leaf_18_edx

Returns Deterministic Address Translation Parameters Information.

leaf_80000000_eax

Extended Function CPUID Information.

leaf_80000000_ebx

Extended Function CPUID Information.

leaf_80000000_ecx

Extended Function CPUID Information.

leaf_80000000_edx

Extended Function CPUID Information.

leaf_80000001_eax

Extended Function CPUID Information.

leaf_80000001_ebx

Extended Function CPUID Information.

leaf_80000001_ecx

Extended Function CPUID Information.

leaf_80000001_edx

Extended Function CPUID Information.

leaf_80000002_eax

Extended Function CPUID Information.

leaf_80000002_ebx

Extended Function CPUID Information.

leaf_80000002_ecx

Extended Function CPUID Information.

leaf_80000002_edx

Extended Function CPUID Information.

leaf_80000003_eax

Extended Function CPUID Information.

leaf_80000003_ebx

Extended Function CPUID Information.

leaf_80000003_ecx

Extended Function CPUID Information.

leaf_80000003_edx

Extended Function CPUID Information.

leaf_80000004_eax

Extended Function CPUID Information.

leaf_80000004_ebx

Extended Function CPUID Information.

leaf_80000004_ecx

Extended Function CPUID Information.

leaf_80000004_edx

Extended Function CPUID Information.

leaf_80000006_eax

Extended Function CPUID Information.

leaf_80000006_ebx

Extended Function CPUID Information.

leaf_80000006_ecx

Extended Function CPUID Information.

leaf_80000006_edx

Extended Function CPUID Information.

leaf_80000007_eax

Extended Function CPUID Information.

leaf_80000007_ebx

Extended Function CPUID Information.

leaf_80000007_ecx

Extended Function CPUID Information.

leaf_80000007_edx

Extended Function CPUID Information.

leaf_80000008_eax

Extended Function CPUID Information.

leaf_80000008_ebx

Extended Function CPUID Information.

leaf_80000008_ecx

Extended Function CPUID Information.

leaf_80000008_edx

Extended Function CPUID Information.