Module pal::cpuid::leaf_07_ecx

Structured Extended Feature Flags Enumeration Leaf.

When CPUID executes with EAX set to 07H and ECX = 0, the processor returns information about the maximum input value for sub-leaves that contain extended feature flags. See Table 3-8. When CPUID executes with EAX set to 07H and the input value of ECX is invalid (see leaf 07H entry in Table 3-8), the processor returns 0 in EAX/EBX/ECX/EDX. In subleaf 0, EAX returns the maximum input value of the highest leaf 7 sub-leaf, and EBX, ECX & EDX contain information of extended feature flags.

Constants

LONG_NAME

NAME

Functions

avx512_vbmi_is_disabled_at_index

Returns whether “AVX512_VBMI” is currently disabled.

avx512_vbmi_is_disabled_in_value

Returns whether “AVX512_VBMI” is disabled in the given register value.

avx512_vbmi_is_enabled_at_index

Returns whether “AVX512_VBMI” is currently enabled.

avx512_vbmi_is_enabled_in_value

Returns whether “AVX512_VBMI” is enabled in the given register value.

avx512_vpopcntdq_is_disabled_at_index

Returns whether “AVX512_VPOPCNTDQ” is currently disabled.

avx512_vpopcntdq_is_disabled_in_value

Returns whether “AVX512_VPOPCNTDQ” is disabled in the given register value.

avx512_vpopcntdq_is_enabled_at_index

Returns whether “AVX512_VPOPCNTDQ” is currently enabled.

avx512_vpopcntdq_is_enabled_in_value

Returns whether “AVX512_VPOPCNTDQ” is enabled in the given register value.

bit_22_is_disabled_at_index

Returns whether “Bit 22” is currently disabled.

bit_22_is_disabled_in_value

Returns whether “Bit 22” is disabled in the given register value.

bit_22_is_enabled_at_index

Returns whether “Bit 22” is currently enabled.

bit_22_is_enabled_in_value

Returns whether “Bit 22” is enabled in the given register value.

cldemote_is_disabled_at_index

Returns whether “CLDEMOTE” is currently disabled.

cldemote_is_disabled_in_value

Returns whether “CLDEMOTE” is disabled in the given register value.

cldemote_is_enabled_at_index

Returns whether “CLDEMOTE” is currently enabled.

cldemote_is_enabled_in_value

Returns whether “CLDEMOTE” is enabled in the given register value.

get_at_index

Returns the current value of the register.

get_mawau_at_index

Returns the current value of “MAWAU”.

get_mawau_from_value

Returns the value of “MAWAU” in the given register value.

gfni_is_disabled_at_index

Returns whether “GFNI” is currently disabled.

gfni_is_disabled_in_value

Returns whether “GFNI” is disabled in the given register value.

gfni_is_enabled_at_index

Returns whether “GFNI” is currently enabled.

gfni_is_enabled_in_value

Returns whether “GFNI” is enabled in the given register value.

movdir64b_is_disabled_at_index

Returns whether “MOVDIR64B” is currently disabled.

movdir64b_is_disabled_in_value

Returns whether “MOVDIR64B” is disabled in the given register value.

movdir64b_is_enabled_at_index

Returns whether “MOVDIR64B” is currently enabled.

movdir64b_is_enabled_in_value

Returns whether “MOVDIR64B” is enabled in the given register value.

movdiri_is_disabled_at_index

Returns whether “MOVDIRI” is currently disabled.

movdiri_is_disabled_in_value

Returns whether “MOVDIRI” is disabled in the given register value.

movdiri_is_enabled_at_index

Returns whether “MOVDIRI” is currently enabled.

movdiri_is_enabled_in_value

Returns whether “MOVDIRI” is enabled in the given register value.

ospke_is_disabled_at_index

Returns whether “OSPKE” is currently disabled.

ospke_is_disabled_in_value

Returns whether “OSPKE” is disabled in the given register value.

ospke_is_enabled_at_index

Returns whether “OSPKE” is currently enabled.

ospke_is_enabled_in_value

Returns whether “OSPKE” is enabled in the given register value.

pku_is_disabled_at_index

Returns whether “protection keys for user-mode pages” is currently disabled.

pku_is_disabled_in_value

Returns whether “protection keys for user-mode pages” is disabled in the given register value.

pku_is_enabled_at_index

Returns whether “protection keys for user-mode pages” is currently enabled.

pku_is_enabled_in_value

Returns whether “protection keys for user-mode pages” is enabled in the given register value.

prefetchwt1_is_disabled_at_index

Returns whether “PREFETCHWT1” is currently disabled.

prefetchwt1_is_disabled_in_value

Returns whether “PREFETCHWT1” is disabled in the given register value.

prefetchwt1_is_enabled_at_index

Returns whether “PREFETCHWT1” is currently enabled.

prefetchwt1_is_enabled_in_value

Returns whether “PREFETCHWT1” is enabled in the given register value.

sgx_lc_is_disabled_at_index

Returns whether “SGX Launch Configuration” is currently disabled.

sgx_lc_is_disabled_in_value

Returns whether “SGX Launch Configuration” is disabled in the given register value.

sgx_lc_is_enabled_at_index

Returns whether “SGX Launch Configuration” is currently enabled.

sgx_lc_is_enabled_in_value

Returns whether “SGX Launch Configuration” is enabled in the given register value.

umip_is_disabled_at_index

Returns whether “user-mode instruction prevention” is currently disabled.

umip_is_disabled_in_value

Returns whether “user-mode instruction prevention” is disabled in the given register value.

umip_is_enabled_at_index

Returns whether “user-mode instruction prevention” is currently enabled.

umip_is_enabled_in_value

Returns whether “user-mode instruction prevention” is enabled in the given register value.

waitpkg_is_disabled_at_index

Returns whether “WAITPKG” is currently disabled.

waitpkg_is_disabled_in_value

Returns whether “WAITPKG” is disabled in the given register value.

waitpkg_is_enabled_at_index

Returns whether “WAITPKG” is currently enabled.

waitpkg_is_enabled_in_value

Returns whether “WAITPKG” is enabled in the given register value.