Module pal::msr::ia32_debug_interface

ia32_debug_interface.

Constants

ADDRESS

LONG_NAME

NAME

Functions

debug_occurred_is_disabled

Returns whether “debug_occurred” is currently disabled.

debug_occurred_is_disabled_in_value

Returns whether “debug_occurred” is disabled in the given register value.

debug_occurred_is_enabled

Returns whether “debug_occurred” is currently enabled.

debug_occurred_is_enabled_in_value

Returns whether “debug_occurred” is enabled in the given register value.

disable_debug_occurred

Disables “debug_occurred”.

disable_debug_occurred_in_value

Disables “debug_occurred” in the given register value, returning the new value.

disable_enable

Disables “enable”.

disable_enable_in_value

Disables “enable” in the given register value, returning the new value.

disable_lock

Disables “lock”.

disable_lock_in_value

Disables “lock” in the given register value, returning the new value.

enable_debug_occurred

Enables “debug_occurred”.

enable_debug_occurred_in_value

Enables “debug_occurred” in the given register value, returning the new value.

enable_enable

Enables “enable”.

enable_enable_in_value

Enables “enable” in the given register value, returning the new value.

enable_is_disabled

Returns whether “enable” is currently disabled.

enable_is_disabled_in_value

Returns whether “enable” is disabled in the given register value.

enable_is_enabled

Returns whether “enable” is currently enabled.

enable_is_enabled_in_value

Returns whether “enable” is enabled in the given register value.

enable_lock

Enables “lock”.

enable_lock_in_value

Enables “lock” in the given register value, returning the new value.

get

Returns the current value of the register.

lock_is_disabled

Returns whether “lock” is currently disabled.

lock_is_disabled_in_value

Returns whether “lock” is disabled in the given register value.

lock_is_enabled

Returns whether “lock” is currently enabled.

lock_is_enabled_in_value

Returns whether “lock” is enabled in the given register value.

set

Sets the value of the register.