Module pal::msr::ia32_mtrr_def_type

ia32_mtrr_def_type.

Constants

ADDRESS

LONG_NAME

NAME

Functions

disable_fixed_range_mtrrs_enable

Disables “fixed_range_mtrr”.

disable_fixed_range_mtrrs_enable_in_value

Disables “fixed_range_mtrr” in the given register value, returning the new value.

disable_mtrr_enable

Disables “mtrr”.

disable_mtrr_enable_in_value

Disables “mtrr” in the given register value, returning the new value.

enable_fixed_range_mtrrs_enable

Enables “fixed_range_mtrr”.

enable_fixed_range_mtrrs_enable_in_value

Enables “fixed_range_mtrr” in the given register value, returning the new value.

enable_mtrr_enable

Enables “mtrr”.

enable_mtrr_enable_in_value

Enables “mtrr” in the given register value, returning the new value.

fixed_range_mtrrs_enable_is_disabled

Returns whether “fixed_range_mtrr” is currently disabled.

fixed_range_mtrrs_enable_is_disabled_in_value

Returns whether “fixed_range_mtrr” is disabled in the given register value.

fixed_range_mtrrs_enable_is_enabled

Returns whether “fixed_range_mtrr” is currently enabled.

fixed_range_mtrrs_enable_is_enabled_in_value

Returns whether “fixed_range_mtrr” is enabled in the given register value.

get

Returns the current value of the register.

get_type

Returns the current value of “type”.

get_type_from_value

Returns the value of “type” in the given register value.

mtrr_enable_is_disabled

Returns whether “mtrr” is currently disabled.

mtrr_enable_is_disabled_in_value

Returns whether “mtrr” is disabled in the given register value.

mtrr_enable_is_enabled

Returns whether “mtrr” is currently enabled.

mtrr_enable_is_enabled_in_value

Returns whether “mtrr” is enabled in the given register value.

set

Sets the value of the register.

set_type

Sets the value of “type”.

set_type_in_value

Sets the value of “type” in the given register value, returning the new value.