Module pal::msr::ia32_package_therm_interrupt

ia32_energy_perf_bias.

Constants

ADDRESS

LONG_NAME

NAME

Functions

disable_pkg_high_temp

Disables “pkg_high_temp”.

disable_pkg_high_temp_in_value

Disables “pkg_high_temp” in the given register value, returning the new value.

disable_pkg_low_temp

Disables “pkg_low_temp”.

disable_pkg_low_temp_in_value

Disables “pkg_low_temp” in the given register value, returning the new value.

disable_pkg_overheat

Disables “pkg_overheat”.

disable_pkg_overheat_in_value

Disables “pkg_overheat” in the given register value, returning the new value.

disable_pkg_power_limit

Disables “pkg_power_limit”.

disable_pkg_power_limit_in_value

Disables “pkg_power_limit” in the given register value, returning the new value.

disable_pkg_prochot

Disables “pkg_prochot”.

disable_pkg_prochot_in_value

Disables “pkg_prochot” in the given register value, returning the new value.

disable_pkg_threshold_1_enable

Disables “pkg_threshold_1_enable”.

disable_pkg_threshold_1_enable_in_value

Disables “pkg_threshold_1_enable” in the given register value, returning the new value.

disable_pkg_threshold_2_enable

Disables “pkg_threshold_2_enable”.

disable_pkg_threshold_2_enable_in_value

Disables “pkg_threshold_2_enable” in the given register value, returning the new value.

enable_pkg_high_temp

Enables “pkg_high_temp”.

enable_pkg_high_temp_in_value

Enables “pkg_high_temp” in the given register value, returning the new value.

enable_pkg_low_temp

Enables “pkg_low_temp”.

enable_pkg_low_temp_in_value

Enables “pkg_low_temp” in the given register value, returning the new value.

enable_pkg_overheat

Enables “pkg_overheat”.

enable_pkg_overheat_in_value

Enables “pkg_overheat” in the given register value, returning the new value.

enable_pkg_power_limit

Enables “pkg_power_limit”.

enable_pkg_power_limit_in_value

Enables “pkg_power_limit” in the given register value, returning the new value.

enable_pkg_prochot

Enables “pkg_prochot”.

enable_pkg_prochot_in_value

Enables “pkg_prochot” in the given register value, returning the new value.

enable_pkg_threshold_1_enable

Enables “pkg_threshold_1_enable”.

enable_pkg_threshold_1_enable_in_value

Enables “pkg_threshold_1_enable” in the given register value, returning the new value.

enable_pkg_threshold_2_enable

Enables “pkg_threshold_2_enable”.

enable_pkg_threshold_2_enable_in_value

Enables “pkg_threshold_2_enable” in the given register value, returning the new value.

get

Returns the current value of the register.

get_pkg_threshold_1_value

Returns the current value of “pkg_threshold_1_value”.

get_pkg_threshold_1_value_from_value

Returns the value of “pkg_threshold_1_value” in the given register value.

get_pkg_threshold_2_value

Returns the current value of “pkg_threshold_2_value”.

get_pkg_threshold_2_value_from_value

Returns the value of “pkg_threshold_2_value” in the given register value.

pkg_high_temp_is_disabled

Returns whether “pkg_high_temp” is currently disabled.

pkg_high_temp_is_disabled_in_value

Returns whether “pkg_high_temp” is disabled in the given register value.

pkg_high_temp_is_enabled

Returns whether “pkg_high_temp” is currently enabled.

pkg_high_temp_is_enabled_in_value

Returns whether “pkg_high_temp” is enabled in the given register value.

pkg_low_temp_is_disabled

Returns whether “pkg_low_temp” is currently disabled.

pkg_low_temp_is_disabled_in_value

Returns whether “pkg_low_temp” is disabled in the given register value.

pkg_low_temp_is_enabled

Returns whether “pkg_low_temp” is currently enabled.

pkg_low_temp_is_enabled_in_value

Returns whether “pkg_low_temp” is enabled in the given register value.

pkg_overheat_is_disabled

Returns whether “pkg_overheat” is currently disabled.

pkg_overheat_is_disabled_in_value

Returns whether “pkg_overheat” is disabled in the given register value.

pkg_overheat_is_enabled

Returns whether “pkg_overheat” is currently enabled.

pkg_overheat_is_enabled_in_value

Returns whether “pkg_overheat” is enabled in the given register value.

pkg_power_limit_is_disabled

Returns whether “pkg_power_limit” is currently disabled.

pkg_power_limit_is_disabled_in_value

Returns whether “pkg_power_limit” is disabled in the given register value.

pkg_power_limit_is_enabled

Returns whether “pkg_power_limit” is currently enabled.

pkg_power_limit_is_enabled_in_value

Returns whether “pkg_power_limit” is enabled in the given register value.

pkg_prochot_is_disabled

Returns whether “pkg_prochot” is currently disabled.

pkg_prochot_is_disabled_in_value

Returns whether “pkg_prochot” is disabled in the given register value.

pkg_prochot_is_enabled

Returns whether “pkg_prochot” is currently enabled.

pkg_prochot_is_enabled_in_value

Returns whether “pkg_prochot” is enabled in the given register value.

pkg_threshold_1_enable_is_disabled

Returns whether “pkg_threshold_1_enable” is currently disabled.

pkg_threshold_1_enable_is_disabled_in_value

Returns whether “pkg_threshold_1_enable” is disabled in the given register value.

pkg_threshold_1_enable_is_enabled

Returns whether “pkg_threshold_1_enable” is currently enabled.

pkg_threshold_1_enable_is_enabled_in_value

Returns whether “pkg_threshold_1_enable” is enabled in the given register value.

pkg_threshold_2_enable_is_disabled

Returns whether “pkg_threshold_2_enable” is currently disabled.

pkg_threshold_2_enable_is_disabled_in_value

Returns whether “pkg_threshold_2_enable” is disabled in the given register value.

pkg_threshold_2_enable_is_enabled

Returns whether “pkg_threshold_2_enable” is currently enabled.

pkg_threshold_2_enable_is_enabled_in_value

Returns whether “pkg_threshold_2_enable” is enabled in the given register value.

set

Sets the value of the register.

set_pkg_threshold_1_value

Sets the value of “pkg_threshold_1_value”.

set_pkg_threshold_1_value_in_value

Sets the value of “pkg_threshold_1_value” in the given register value, returning the new value.

set_pkg_threshold_2_value

Sets the value of “pkg_threshold_2_value”.

set_pkg_threshold_2_value_in_value

Sets the value of “pkg_threshold_2_value” in the given register value, returning the new value.