Module pal::msr::ia32_perf_global_ovf_ctrl

ia32_perf_global_ovf_ctrl.

Constants

ADDRESS

LONG_NAME

NAME

Functions

clear_condchgd_is_disabled

Returns whether “clear_condchgd” is currently disabled.

clear_condchgd_is_disabled_in_value

Returns whether “clear_condchgd” is disabled in the given register value.

clear_condchgd_is_enabled

Returns whether “clear_condchgd” is currently enabled.

clear_condchgd_is_enabled_in_value

Returns whether “clear_condchgd” is enabled in the given register value.

clear_ovf_fixed_ctr0_is_disabled

Returns whether “clear_ovf_fixed_ctr0” is currently disabled.

clear_ovf_fixed_ctr0_is_disabled_in_value

Returns whether “clear_ovf_fixed_ctr0” is disabled in the given register value.

clear_ovf_fixed_ctr0_is_enabled

Returns whether “clear_ovf_fixed_ctr0” is currently enabled.

clear_ovf_fixed_ctr0_is_enabled_in_value

Returns whether “clear_ovf_fixed_ctr0” is enabled in the given register value.

clear_ovf_fixed_ctr1_is_disabled

Returns whether “clear_ovf_fixed_ctr1” is currently disabled.

clear_ovf_fixed_ctr1_is_disabled_in_value

Returns whether “clear_ovf_fixed_ctr1” is disabled in the given register value.

clear_ovf_fixed_ctr1_is_enabled

Returns whether “clear_ovf_fixed_ctr1” is currently enabled.

clear_ovf_fixed_ctr1_is_enabled_in_value

Returns whether “clear_ovf_fixed_ctr1” is enabled in the given register value.

clear_ovf_fixed_ctr2_is_disabled

Returns whether “clear_ovf_fixed_ctr2” is currently disabled.

clear_ovf_fixed_ctr2_is_disabled_in_value

Returns whether “clear_ovf_fixed_ctr2” is disabled in the given register value.

clear_ovf_fixed_ctr2_is_enabled

Returns whether “clear_ovf_fixed_ctr2” is currently enabled.

clear_ovf_fixed_ctr2_is_enabled_in_value

Returns whether “clear_ovf_fixed_ctr2” is enabled in the given register value.

clear_ovf_pmc0_is_disabled

Returns whether “clear_ovf_pmc0” is currently disabled.

clear_ovf_pmc0_is_disabled_in_value

Returns whether “clear_ovf_pmc0” is disabled in the given register value.

clear_ovf_pmc0_is_enabled

Returns whether “clear_ovf_pmc0” is currently enabled.

clear_ovf_pmc0_is_enabled_in_value

Returns whether “clear_ovf_pmc0” is enabled in the given register value.

clear_ovf_pmc1_is_disabled

Returns whether “clear_ovf_pmc1” is currently disabled.

clear_ovf_pmc1_is_disabled_in_value

Returns whether “clear_ovf_pmc1” is disabled in the given register value.

clear_ovf_pmc1_is_enabled

Returns whether “clear_ovf_pmc1” is currently enabled.

clear_ovf_pmc1_is_enabled_in_value

Returns whether “clear_ovf_pmc1” is enabled in the given register value.

clear_ovf_pmc2_is_disabled

Returns whether “clear_ovf_pmc2” is currently disabled.

clear_ovf_pmc2_is_disabled_in_value

Returns whether “clear_ovf_pmc2” is disabled in the given register value.

clear_ovf_pmc2_is_enabled

Returns whether “clear_ovf_pmc2” is currently enabled.

clear_ovf_pmc2_is_enabled_in_value

Returns whether “clear_ovf_pmc2” is enabled in the given register value.

clear_ovf_uncore_is_disabled

Returns whether “clear_ovf_uncore” is currently disabled.

clear_ovf_uncore_is_disabled_in_value

Returns whether “clear_ovf_uncore” is disabled in the given register value.

clear_ovf_uncore_is_enabled

Returns whether “clear_ovf_uncore” is currently enabled.

clear_ovf_uncore_is_enabled_in_value

Returns whether “clear_ovf_uncore” is enabled in the given register value.

clear_ovfbuf_is_disabled

Returns whether “clear_ovfbuf” is currently disabled.

clear_ovfbuf_is_disabled_in_value

Returns whether “clear_ovfbuf” is disabled in the given register value.

clear_ovfbuf_is_enabled

Returns whether “clear_ovfbuf” is currently enabled.

clear_ovfbuf_is_enabled_in_value

Returns whether “clear_ovfbuf” is enabled in the given register value.

clear_trace_topa_pmi_is_disabled

Returns whether “clear_trace_topa_pmi” is currently disabled.

clear_trace_topa_pmi_is_disabled_in_value

Returns whether “clear_trace_topa_pmi” is disabled in the given register value.

clear_trace_topa_pmi_is_enabled

Returns whether “clear_trace_topa_pmi” is currently enabled.

clear_trace_topa_pmi_is_enabled_in_value

Returns whether “clear_trace_topa_pmi” is enabled in the given register value.

ctr_frz_is_disabled

Returns whether “ctr_frz” is currently disabled.

ctr_frz_is_disabled_in_value

Returns whether “ctr_frz” is disabled in the given register value.

ctr_frz_is_enabled

Returns whether “ctr_frz” is currently enabled.

ctr_frz_is_enabled_in_value

Returns whether “ctr_frz” is enabled in the given register value.

disable_clear_condchgd

Disables “clear_condchgd”.

disable_clear_condchgd_in_value

Disables “clear_condchgd” in the given register value, returning the new value.

disable_clear_ovf_fixed_ctr0

Disables “clear_ovf_fixed_ctr0”.

disable_clear_ovf_fixed_ctr0_in_value

Disables “clear_ovf_fixed_ctr0” in the given register value, returning the new value.

disable_clear_ovf_fixed_ctr1

Disables “clear_ovf_fixed_ctr1”.

disable_clear_ovf_fixed_ctr1_in_value

Disables “clear_ovf_fixed_ctr1” in the given register value, returning the new value.

disable_clear_ovf_fixed_ctr2

Disables “clear_ovf_fixed_ctr2”.

disable_clear_ovf_fixed_ctr2_in_value

Disables “clear_ovf_fixed_ctr2” in the given register value, returning the new value.

disable_clear_ovf_pmc0

Disables “clear_ovf_pmc0”.

disable_clear_ovf_pmc0_in_value

Disables “clear_ovf_pmc0” in the given register value, returning the new value.

disable_clear_ovf_pmc1

Disables “clear_ovf_pmc1”.

disable_clear_ovf_pmc1_in_value

Disables “clear_ovf_pmc1” in the given register value, returning the new value.

disable_clear_ovf_pmc2

Disables “clear_ovf_pmc2”.

disable_clear_ovf_pmc2_in_value

Disables “clear_ovf_pmc2” in the given register value, returning the new value.

disable_clear_ovf_uncore

Disables “clear_ovf_uncore”.

disable_clear_ovf_uncore_in_value

Disables “clear_ovf_uncore” in the given register value, returning the new value.

disable_clear_ovfbuf

Disables “clear_ovfbuf”.

disable_clear_ovfbuf_in_value

Disables “clear_ovfbuf” in the given register value, returning the new value.

disable_clear_trace_topa_pmi

Disables “clear_trace_topa_pmi”.

disable_clear_trace_topa_pmi_in_value

Disables “clear_trace_topa_pmi” in the given register value, returning the new value.

disable_ctr_frz

Disables “ctr_frz”.

disable_ctr_frz_in_value

Disables “ctr_frz” in the given register value, returning the new value.

disable_lbr_frz

Disables “lbr_frz”.

disable_lbr_frz_in_value

Disables “lbr_frz” in the given register value, returning the new value.

enable_clear_condchgd

Enables “clear_condchgd”.

enable_clear_condchgd_in_value

Enables “clear_condchgd” in the given register value, returning the new value.

enable_clear_ovf_fixed_ctr0

Enables “clear_ovf_fixed_ctr0”.

enable_clear_ovf_fixed_ctr0_in_value

Enables “clear_ovf_fixed_ctr0” in the given register value, returning the new value.

enable_clear_ovf_fixed_ctr1

Enables “clear_ovf_fixed_ctr1”.

enable_clear_ovf_fixed_ctr1_in_value

Enables “clear_ovf_fixed_ctr1” in the given register value, returning the new value.

enable_clear_ovf_fixed_ctr2

Enables “clear_ovf_fixed_ctr2”.

enable_clear_ovf_fixed_ctr2_in_value

Enables “clear_ovf_fixed_ctr2” in the given register value, returning the new value.

enable_clear_ovf_pmc0

Enables “clear_ovf_pmc0”.

enable_clear_ovf_pmc0_in_value

Enables “clear_ovf_pmc0” in the given register value, returning the new value.

enable_clear_ovf_pmc1

Enables “clear_ovf_pmc1”.

enable_clear_ovf_pmc1_in_value

Enables “clear_ovf_pmc1” in the given register value, returning the new value.

enable_clear_ovf_pmc2

Enables “clear_ovf_pmc2”.

enable_clear_ovf_pmc2_in_value

Enables “clear_ovf_pmc2” in the given register value, returning the new value.

enable_clear_ovf_uncore

Enables “clear_ovf_uncore”.

enable_clear_ovf_uncore_in_value

Enables “clear_ovf_uncore” in the given register value, returning the new value.

enable_clear_ovfbuf

Enables “clear_ovfbuf”.

enable_clear_ovfbuf_in_value

Enables “clear_ovfbuf” in the given register value, returning the new value.

enable_clear_trace_topa_pmi

Enables “clear_trace_topa_pmi”.

enable_clear_trace_topa_pmi_in_value

Enables “clear_trace_topa_pmi” in the given register value, returning the new value.

enable_ctr_frz

Enables “ctr_frz”.

enable_ctr_frz_in_value

Enables “ctr_frz” in the given register value, returning the new value.

enable_lbr_frz

Enables “lbr_frz”.

enable_lbr_frz_in_value

Enables “lbr_frz” in the given register value, returning the new value.

get

Returns the current value of the register.

lbr_frz_is_disabled

Returns whether “lbr_frz” is currently disabled.

lbr_frz_is_disabled_in_value

Returns whether “lbr_frz” is disabled in the given register value.

lbr_frz_is_enabled

Returns whether “lbr_frz” is currently enabled.

lbr_frz_is_enabled_in_value

Returns whether “lbr_frz” is enabled in the given register value.

set

Sets the value of the register.