Module pal::msr::ia32_perfevtsel0

ia32_perfevtsel0.

Constants

ADDRESS

LONG_NAME

NAME

Functions

anythread_is_disabled

Returns whether “anythread” is currently disabled.

anythread_is_disabled_in_value

Returns whether “anythread” is disabled in the given register value.

anythread_is_enabled

Returns whether “anythread” is currently enabled.

anythread_is_enabled_in_value

Returns whether “anythread” is enabled in the given register value.

disable_anythread

Disables “anythread”.

disable_anythread_in_value

Disables “anythread” in the given register value, returning the new value.

disable_edge

Disables “edge”.

disable_edge_in_value

Disables “edge” in the given register value, returning the new value.

disable_en

Disables “en”.

disable_en_in_value

Disables “en” in the given register value, returning the new value.

disable_interrupt

Disables “interrupt”.

disable_interrupt_in_value

Disables “interrupt” in the given register value, returning the new value.

disable_inv

Disables “inv”.

disable_inv_in_value

Disables “inv” in the given register value, returning the new value.

disable_os

Disables “os”.

disable_os_in_value

Disables “os” in the given register value, returning the new value.

disable_pc

Disables “pc”.

disable_pc_in_value

Disables “pc” in the given register value, returning the new value.

disable_usr

Disables “usr”.

disable_usr_in_value

Disables “usr” in the given register value, returning the new value.

edge_is_disabled

Returns whether “edge” is currently disabled.

edge_is_disabled_in_value

Returns whether “edge” is disabled in the given register value.

edge_is_enabled

Returns whether “edge” is currently enabled.

edge_is_enabled_in_value

Returns whether “edge” is enabled in the given register value.

en_is_disabled

Returns whether “en” is currently disabled.

en_is_disabled_in_value

Returns whether “en” is disabled in the given register value.

en_is_enabled

Returns whether “en” is currently enabled.

en_is_enabled_in_value

Returns whether “en” is enabled in the given register value.

enable_anythread

Enables “anythread”.

enable_anythread_in_value

Enables “anythread” in the given register value, returning the new value.

enable_edge

Enables “edge”.

enable_edge_in_value

Enables “edge” in the given register value, returning the new value.

enable_en

Enables “en”.

enable_en_in_value

Enables “en” in the given register value, returning the new value.

enable_interrupt

Enables “interrupt”.

enable_interrupt_in_value

Enables “interrupt” in the given register value, returning the new value.

enable_inv

Enables “inv”.

enable_inv_in_value

Enables “inv” in the given register value, returning the new value.

enable_os

Enables “os”.

enable_os_in_value

Enables “os” in the given register value, returning the new value.

enable_pc

Enables “pc”.

enable_pc_in_value

Enables “pc” in the given register value, returning the new value.

enable_usr

Enables “usr”.

enable_usr_in_value

Enables “usr” in the given register value, returning the new value.

get

Returns the current value of the register.

get_cmask

Returns the current value of “cmask”.

get_cmask_from_value

Returns the value of “cmask” in the given register value.

get_event_select

Returns the current value of “event_select”.

get_event_select_from_value

Returns the value of “event_select” in the given register value.

get_umask

Returns the current value of “umask”.

get_umask_from_value

Returns the value of “umask” in the given register value.

interrupt_is_disabled

Returns whether “interrupt” is currently disabled.

interrupt_is_disabled_in_value

Returns whether “interrupt” is disabled in the given register value.

interrupt_is_enabled

Returns whether “interrupt” is currently enabled.

interrupt_is_enabled_in_value

Returns whether “interrupt” is enabled in the given register value.

inv_is_disabled

Returns whether “inv” is currently disabled.

inv_is_disabled_in_value

Returns whether “inv” is disabled in the given register value.

inv_is_enabled

Returns whether “inv” is currently enabled.

inv_is_enabled_in_value

Returns whether “inv” is enabled in the given register value.

os_is_disabled

Returns whether “os” is currently disabled.

os_is_disabled_in_value

Returns whether “os” is disabled in the given register value.

os_is_enabled

Returns whether “os” is currently enabled.

os_is_enabled_in_value

Returns whether “os” is enabled in the given register value.

pc_is_disabled

Returns whether “pc” is currently disabled.

pc_is_disabled_in_value

Returns whether “pc” is disabled in the given register value.

pc_is_enabled

Returns whether “pc” is currently enabled.

pc_is_enabled_in_value

Returns whether “pc” is enabled in the given register value.

set

Sets the value of the register.

set_cmask

Sets the value of “cmask”.

set_cmask_in_value

Sets the value of “cmask” in the given register value, returning the new value.

set_event_select

Sets the value of “event_select”.

set_event_select_in_value

Sets the value of “event_select” in the given register value, returning the new value.

set_umask

Sets the value of “umask”.

set_umask_in_value

Sets the value of “umask” in the given register value, returning the new value.

usr_is_disabled

Returns whether “usr” is currently disabled.

usr_is_disabled_in_value

Returns whether “usr” is disabled in the given register value.

usr_is_enabled

Returns whether “usr” is currently enabled.

usr_is_enabled_in_value

Returns whether “usr” is enabled in the given register value.