Module pal::msr::ia32_rtit_ctl

ia32_rtit_ctl.

Constants

ADDRESS

LONG_NAME

NAME

Functions

branchen_is_disabled

Returns whether “branchen” is currently disabled.

branchen_is_disabled_in_value

Returns whether “branchen” is disabled in the given register value.

branchen_is_enabled

Returns whether “branchen” is currently enabled.

branchen_is_enabled_in_value

Returns whether “branchen” is enabled in the given register value.

cr3_filter_is_disabled

Returns whether “cr3_filter” is currently disabled.

cr3_filter_is_disabled_in_value

Returns whether “cr3_filter” is disabled in the given register value.

cr3_filter_is_enabled

Returns whether “cr3_filter” is currently enabled.

cr3_filter_is_enabled_in_value

Returns whether “cr3_filter” is enabled in the given register value.

cycen_is_disabled

Returns whether “cycen” is currently disabled.

cycen_is_disabled_in_value

Returns whether “cycen” is disabled in the given register value.

cycen_is_enabled

Returns whether “cycen” is currently enabled.

cycen_is_enabled_in_value

Returns whether “cycen” is enabled in the given register value.

disable_branchen

Disables “branchen”.

disable_branchen_in_value

Disables “branchen” in the given register value, returning the new value.

disable_cr3_filter

Disables “cr3_filter”.

disable_cr3_filter_in_value

Disables “cr3_filter” in the given register value, returning the new value.

disable_cycen

Disables “cycen”.

disable_cycen_in_value

Disables “cycen” in the given register value, returning the new value.

disable_disretc

Disables “disretc”.

disable_disretc_in_value

Disables “disretc” in the given register value, returning the new value.

disable_fabricen

Disables “fabricen”.

disable_fabricen_in_value

Disables “fabricen” in the given register value, returning the new value.

disable_mtcen

Disables “mtcen”.

disable_mtcen_in_value

Disables “mtcen” in the given register value, returning the new value.

disable_os

Disables “os”.

disable_os_in_value

Disables “os” in the given register value, returning the new value.

disable_topa

Disables “topa”.

disable_topa_in_value

Disables “topa” in the given register value, returning the new value.

disable_traceen

Disables “traceen”.

disable_traceen_in_value

Disables “traceen” in the given register value, returning the new value.

disable_tscen

Disables “tscen”.

disable_tscen_in_value

Disables “tscen” in the given register value, returning the new value.

disable_user

Disables “user”.

disable_user_in_value

Disables “user” in the given register value, returning the new value.

disretc_is_disabled

Returns whether “disretc” is currently disabled.

disretc_is_disabled_in_value

Returns whether “disretc” is disabled in the given register value.

disretc_is_enabled

Returns whether “disretc” is currently enabled.

disretc_is_enabled_in_value

Returns whether “disretc” is enabled in the given register value.

enable_branchen

Enables “branchen”.

enable_branchen_in_value

Enables “branchen” in the given register value, returning the new value.

enable_cr3_filter

Enables “cr3_filter”.

enable_cr3_filter_in_value

Enables “cr3_filter” in the given register value, returning the new value.

enable_cycen

Enables “cycen”.

enable_cycen_in_value

Enables “cycen” in the given register value, returning the new value.

enable_disretc

Enables “disretc”.

enable_disretc_in_value

Enables “disretc” in the given register value, returning the new value.

enable_fabricen

Enables “fabricen”.

enable_fabricen_in_value

Enables “fabricen” in the given register value, returning the new value.

enable_mtcen

Enables “mtcen”.

enable_mtcen_in_value

Enables “mtcen” in the given register value, returning the new value.

enable_os

Enables “os”.

enable_os_in_value

Enables “os” in the given register value, returning the new value.

enable_topa

Enables “topa”.

enable_topa_in_value

Enables “topa” in the given register value, returning the new value.

enable_traceen

Enables “traceen”.

enable_traceen_in_value

Enables “traceen” in the given register value, returning the new value.

enable_tscen

Enables “tscen”.

enable_tscen_in_value

Enables “tscen” in the given register value, returning the new value.

enable_user

Enables “user”.

enable_user_in_value

Enables “user” in the given register value, returning the new value.

fabricen_is_disabled

Returns whether “fabricen” is currently disabled.

fabricen_is_disabled_in_value

Returns whether “fabricen” is disabled in the given register value.

fabricen_is_enabled

Returns whether “fabricen” is currently enabled.

fabricen_is_enabled_in_value

Returns whether “fabricen” is enabled in the given register value.

get

Returns the current value of the register.

get_addr0_cfg

Returns the current value of “addr0_cfg”.

get_addr0_cfg_from_value

Returns the value of “addr0_cfg” in the given register value.

get_addr1_cfg

Returns the current value of “addr1_cfg”.

get_addr1_cfg_from_value

Returns the value of “addr1_cfg” in the given register value.

get_addr2_cfg

Returns the current value of “addr2_cfg”.

get_addr2_cfg_from_value

Returns the value of “addr2_cfg” in the given register value.

get_addr3_cfg

Returns the current value of “addr3_cfg”.

get_addr3_cfg_from_value

Returns the value of “addr3_cfg” in the given register value.

get_cycthresh

Returns the current value of “cycthresh”.

get_cycthresh_from_value

Returns the value of “cycthresh” in the given register value.

get_mtcfreq

Returns the current value of “mtcfreq”.

get_mtcfreq_from_value

Returns the value of “mtcfreq” in the given register value.

get_psbfreq

Returns the current value of “psbfreq”.

get_psbfreq_from_value

Returns the value of “psbfreq” in the given register value.

mtcen_is_disabled

Returns whether “mtcen” is currently disabled.

mtcen_is_disabled_in_value

Returns whether “mtcen” is disabled in the given register value.

mtcen_is_enabled

Returns whether “mtcen” is currently enabled.

mtcen_is_enabled_in_value

Returns whether “mtcen” is enabled in the given register value.

os_is_disabled

Returns whether “os” is currently disabled.

os_is_disabled_in_value

Returns whether “os” is disabled in the given register value.

os_is_enabled

Returns whether “os” is currently enabled.

os_is_enabled_in_value

Returns whether “os” is enabled in the given register value.

set

Sets the value of the register.

set_addr0_cfg

Sets the value of “addr0_cfg”.

set_addr0_cfg_in_value

Sets the value of “addr0_cfg” in the given register value, returning the new value.

set_addr1_cfg

Sets the value of “addr1_cfg”.

set_addr1_cfg_in_value

Sets the value of “addr1_cfg” in the given register value, returning the new value.

set_addr2_cfg

Sets the value of “addr2_cfg”.

set_addr2_cfg_in_value

Sets the value of “addr2_cfg” in the given register value, returning the new value.

set_addr3_cfg

Sets the value of “addr3_cfg”.

set_addr3_cfg_in_value

Sets the value of “addr3_cfg” in the given register value, returning the new value.

set_cycthresh

Sets the value of “cycthresh”.

set_cycthresh_in_value

Sets the value of “cycthresh” in the given register value, returning the new value.

set_mtcfreq

Sets the value of “mtcfreq”.

set_mtcfreq_in_value

Sets the value of “mtcfreq” in the given register value, returning the new value.

set_psbfreq

Sets the value of “psbfreq”.

set_psbfreq_in_value

Sets the value of “psbfreq” in the given register value, returning the new value.

topa_is_disabled

Returns whether “topa” is currently disabled.

topa_is_disabled_in_value

Returns whether “topa” is disabled in the given register value.

topa_is_enabled

Returns whether “topa” is currently enabled.

topa_is_enabled_in_value

Returns whether “topa” is enabled in the given register value.

traceen_is_disabled

Returns whether “traceen” is currently disabled.

traceen_is_disabled_in_value

Returns whether “traceen” is disabled in the given register value.

traceen_is_enabled

Returns whether “traceen” is currently enabled.

traceen_is_enabled_in_value

Returns whether “traceen” is enabled in the given register value.

tscen_is_disabled

Returns whether “tscen” is currently disabled.

tscen_is_disabled_in_value

Returns whether “tscen” is disabled in the given register value.

tscen_is_enabled

Returns whether “tscen” is currently enabled.

tscen_is_enabled_in_value

Returns whether “tscen” is enabled in the given register value.

user_is_disabled

Returns whether “user” is currently disabled.

user_is_disabled_in_value

Returns whether “user” is disabled in the given register value.

user_is_enabled

Returns whether “user” is currently enabled.

user_is_enabled_in_value

Returns whether “user” is enabled in the given register value.