Module pal::msr::ia32_rtit_status

ia32_rtit_status.

Constants

ADDRESS

LONG_NAME

NAME

Functions

contexen_is_disabled

Returns whether “contexen” is currently disabled.

contexen_is_disabled_in_value

Returns whether “contexen” is disabled in the given register value.

contexen_is_enabled

Returns whether “contexen” is currently enabled.

contexen_is_enabled_in_value

Returns whether “contexen” is enabled in the given register value.

disable_contexen

Disables “contexen”.

disable_contexen_in_value

Disables “contexen” in the given register value, returning the new value.

disable_error

Disables “error”.

disable_error_in_value

Disables “error” in the given register value, returning the new value.

disable_filteren

Disables “filteren”.

disable_filteren_in_value

Disables “filteren” in the given register value, returning the new value.

disable_stopped

Disables “stopped”.

disable_stopped_in_value

Disables “stopped” in the given register value, returning the new value.

disable_triggeren

Disables “triggeren”.

disable_triggeren_in_value

Disables “triggeren” in the given register value, returning the new value.

enable_contexen

Enables “contexen”.

enable_contexen_in_value

Enables “contexen” in the given register value, returning the new value.

enable_error

Enables “error”.

enable_error_in_value

Enables “error” in the given register value, returning the new value.

enable_filteren

Enables “filteren”.

enable_filteren_in_value

Enables “filteren” in the given register value, returning the new value.

enable_stopped

Enables “stopped”.

enable_stopped_in_value

Enables “stopped” in the given register value, returning the new value.

enable_triggeren

Enables “triggeren”.

enable_triggeren_in_value

Enables “triggeren” in the given register value, returning the new value.

error_is_disabled

Returns whether “error” is currently disabled.

error_is_disabled_in_value

Returns whether “error” is disabled in the given register value.

error_is_enabled

Returns whether “error” is currently enabled.

error_is_enabled_in_value

Returns whether “error” is enabled in the given register value.

filteren_is_disabled

Returns whether “filteren” is currently disabled.

filteren_is_disabled_in_value

Returns whether “filteren” is disabled in the given register value.

filteren_is_enabled

Returns whether “filteren” is currently enabled.

filteren_is_enabled_in_value

Returns whether “filteren” is enabled in the given register value.

get

Returns the current value of the register.

get_packetbytecnt

Returns the current value of “packetbytecnt”.

get_packetbytecnt_from_value

Returns the value of “packetbytecnt” in the given register value.

set

Sets the value of the register.

set_packetbytecnt

Sets the value of “packetbytecnt”.

set_packetbytecnt_in_value

Sets the value of “packetbytecnt” in the given register value, returning the new value.

stopped_is_disabled

Returns whether “stopped” is currently disabled.

stopped_is_disabled_in_value

Returns whether “stopped” is disabled in the given register value.

stopped_is_enabled

Returns whether “stopped” is currently enabled.

stopped_is_enabled_in_value

Returns whether “stopped” is enabled in the given register value.

triggeren_is_disabled

Returns whether “triggeren” is currently disabled.

triggeren_is_disabled_in_value

Returns whether “triggeren” is disabled in the given register value.

triggeren_is_enabled

Returns whether “triggeren” is currently enabled.

triggeren_is_enabled_in_value

Returns whether “triggeren” is enabled in the given register value.