Module pal::msr::ia32_therm_interrupt

ia32_therm_interrupt.

Constants

ADDRESS

LONG_NAME

NAME

Functions

crit_temp_is_disabled

Returns whether “crit_temp” is currently disabled.

crit_temp_is_disabled_in_value

Returns whether “crit_temp” is disabled in the given register value.

crit_temp_is_enabled

Returns whether “crit_temp” is currently enabled.

crit_temp_is_enabled_in_value

Returns whether “crit_temp” is enabled in the given register value.

disable_crit_temp

Disables “crit_temp”.

disable_crit_temp_in_value

Disables “crit_temp” in the given register value, returning the new value.

disable_forcepr

Disables “forcepr”.

disable_forcepr_in_value

Disables “forcepr” in the given register value, returning the new value.

disable_high_temp

Disables “high_temp”.

disable_high_temp_in_value

Disables “high_temp” in the given register value, returning the new value.

disable_low_temp

Disables “low_temp”.

disable_low_temp_in_value

Disables “low_temp” in the given register value, returning the new value.

disable_power_limit

Disables “power_limit”.

disable_power_limit_in_value

Disables “power_limit” in the given register value, returning the new value.

disable_prochot

Disables “prochot”.

disable_prochot_in_value

Disables “prochot” in the given register value, returning the new value.

disable_threshold_1_enable

Disables “threshold_1_enable”.

disable_threshold_1_enable_in_value

Disables “threshold_1_enable” in the given register value, returning the new value.

disable_threshold_2_enable

Disables “threshold_2_enable”.

disable_threshold_2_enable_in_value

Disables “threshold_2_enable” in the given register value, returning the new value.

enable_crit_temp

Enables “crit_temp”.

enable_crit_temp_in_value

Enables “crit_temp” in the given register value, returning the new value.

enable_forcepr

Enables “forcepr”.

enable_forcepr_in_value

Enables “forcepr” in the given register value, returning the new value.

enable_high_temp

Enables “high_temp”.

enable_high_temp_in_value

Enables “high_temp” in the given register value, returning the new value.

enable_low_temp

Enables “low_temp”.

enable_low_temp_in_value

Enables “low_temp” in the given register value, returning the new value.

enable_power_limit

Enables “power_limit”.

enable_power_limit_in_value

Enables “power_limit” in the given register value, returning the new value.

enable_prochot

Enables “prochot”.

enable_prochot_in_value

Enables “prochot” in the given register value, returning the new value.

enable_threshold_1_enable

Enables “threshold_1_enable”.

enable_threshold_1_enable_in_value

Enables “threshold_1_enable” in the given register value, returning the new value.

enable_threshold_2_enable

Enables “threshold_2_enable”.

enable_threshold_2_enable_in_value

Enables “threshold_2_enable” in the given register value, returning the new value.

forcepr_is_disabled

Returns whether “forcepr” is currently disabled.

forcepr_is_disabled_in_value

Returns whether “forcepr” is disabled in the given register value.

forcepr_is_enabled

Returns whether “forcepr” is currently enabled.

forcepr_is_enabled_in_value

Returns whether “forcepr” is enabled in the given register value.

get

Returns the current value of the register.

get_threshold_1_value

Returns the current value of “threshold_1_value”.

get_threshold_1_value_from_value

Returns the value of “threshold_1_value” in the given register value.

get_threshold_2_value

Returns the current value of “threshold_2_value”.

get_threshold_2_value_from_value

Returns the value of “threshold_2_value” in the given register value.

high_temp_is_disabled

Returns whether “high_temp” is currently disabled.

high_temp_is_disabled_in_value

Returns whether “high_temp” is disabled in the given register value.

high_temp_is_enabled

Returns whether “high_temp” is currently enabled.

high_temp_is_enabled_in_value

Returns whether “high_temp” is enabled in the given register value.

low_temp_is_disabled

Returns whether “low_temp” is currently disabled.

low_temp_is_disabled_in_value

Returns whether “low_temp” is disabled in the given register value.

low_temp_is_enabled

Returns whether “low_temp” is currently enabled.

low_temp_is_enabled_in_value

Returns whether “low_temp” is enabled in the given register value.

power_limit_is_disabled

Returns whether “power_limit” is currently disabled.

power_limit_is_disabled_in_value

Returns whether “power_limit” is disabled in the given register value.

power_limit_is_enabled

Returns whether “power_limit” is currently enabled.

power_limit_is_enabled_in_value

Returns whether “power_limit” is enabled in the given register value.

prochot_is_disabled

Returns whether “prochot” is currently disabled.

prochot_is_disabled_in_value

Returns whether “prochot” is disabled in the given register value.

prochot_is_enabled

Returns whether “prochot” is currently enabled.

prochot_is_enabled_in_value

Returns whether “prochot” is enabled in the given register value.

set

Sets the value of the register.

set_threshold_1_value

Sets the value of “threshold_1_value”.

set_threshold_1_value_in_value

Sets the value of “threshold_1_value” in the given register value, returning the new value.

set_threshold_2_value

Sets the value of “threshold_2_value”.

set_threshold_2_value_in_value

Sets the value of “threshold_2_value” in the given register value, returning the new value.

threshold_1_enable_is_disabled

Returns whether “threshold_1_enable” is currently disabled.

threshold_1_enable_is_disabled_in_value

Returns whether “threshold_1_enable” is disabled in the given register value.

threshold_1_enable_is_enabled

Returns whether “threshold_1_enable” is currently enabled.

threshold_1_enable_is_enabled_in_value

Returns whether “threshold_1_enable” is enabled in the given register value.

threshold_2_enable_is_disabled

Returns whether “threshold_2_enable” is currently disabled.

threshold_2_enable_is_disabled_in_value

Returns whether “threshold_2_enable” is disabled in the given register value.

threshold_2_enable_is_enabled

Returns whether “threshold_2_enable” is currently enabled.

threshold_2_enable_is_enabled_in_value

Returns whether “threshold_2_enable” is enabled in the given register value.