Module pal::vmcs::guest_ia32_debugctl

Guest IA32_DEBUGCTL (full).

Constants

ENCODING

LONG_NAME

NAME

Functions

bt_off_os_is_disabled

Returns whether “bt_off_os” is currently disabled.

bt_off_os_is_disabled_in_value

Returns whether “bt_off_os” is disabled in the given register value.

bt_off_os_is_enabled

Returns whether “bt_off_os” is currently enabled.

bt_off_os_is_enabled_in_value

Returns whether “bt_off_os” is enabled in the given register value.

bt_off_user_is_disabled

Returns whether “bt_off_user” is currently disabled.

bt_off_user_is_disabled_in_value

Returns whether “bt_off_user” is disabled in the given register value.

bt_off_user_is_enabled

Returns whether “bt_off_user” is currently enabled.

bt_off_user_is_enabled_in_value

Returns whether “bt_off_user” is enabled in the given register value.

btf_is_disabled

Returns whether “btf” is currently disabled.

btf_is_disabled_in_value

Returns whether “btf” is disabled in the given register value.

btf_is_enabled

Returns whether “btf” is currently enabled.

btf_is_enabled_in_value

Returns whether “btf” is enabled in the given register value.

btint_is_disabled

Returns whether “btint” is currently disabled.

btint_is_disabled_in_value

Returns whether “btint” is disabled in the given register value.

btint_is_enabled

Returns whether “btint” is currently enabled.

btint_is_enabled_in_value

Returns whether “btint” is enabled in the given register value.

bts_is_disabled

Returns whether “bts” is currently disabled.

bts_is_disabled_in_value

Returns whether “bts” is disabled in the given register value.

bts_is_enabled

Returns whether “bts” is currently enabled.

bts_is_enabled_in_value

Returns whether “bts” is enabled in the given register value.

disable_bt_off_os

Disables “bt_off_os”.

disable_bt_off_os_in_value

Disables “bt_off_os” in the given register value, returning the new value.

disable_bt_off_user

Disables “bt_off_user”.

disable_bt_off_user_in_value

Disables “bt_off_user” in the given register value, returning the new value.

disable_btf

Disables “btf”.

disable_btf_in_value

Disables “btf” in the given register value, returning the new value.

disable_btint

Disables “btint”.

disable_btint_in_value

Disables “btint” in the given register value, returning the new value.

disable_bts

Disables “bts”.

disable_bts_in_value

Disables “bts” in the given register value, returning the new value.

disable_enable_uncore_pmi

Disables “enable_uncore_pmi”.

disable_enable_uncore_pmi_in_value

Disables “enable_uncore_pmi” in the given register value, returning the new value.

disable_freeze_lbrs_on_pmi

Disables “freeze_lbrs_on_pmi”.

disable_freeze_lbrs_on_pmi_in_value

Disables “freeze_lbrs_on_pmi” in the given register value, returning the new value.

disable_freeze_perfmon_on_pmi

Disables “freeze_perfmon_on_pmi”.

disable_freeze_perfmon_on_pmi_in_value

Disables “freeze_perfmon_on_pmi” in the given register value, returning the new value.

disable_freeze_while_smm

Disables “freeze_while_smm”.

disable_freeze_while_smm_in_value

Disables “freeze_while_smm” in the given register value, returning the new value.

disable_lbr

Disables “lbr”.

disable_lbr_in_value

Disables “lbr” in the given register value, returning the new value.

disable_rtm_debug

Disables “rtm_debug”.

disable_rtm_debug_in_value

Disables “rtm_debug” in the given register value, returning the new value.

disable_tr

Disables “tr”.

disable_tr_in_value

Disables “tr” in the given register value, returning the new value.

enable_bt_off_os

Enables “bt_off_os”.

enable_bt_off_os_in_value

Enables “bt_off_os” in the given register value, returning the new value.

enable_bt_off_user

Enables “bt_off_user”.

enable_bt_off_user_in_value

Enables “bt_off_user” in the given register value, returning the new value.

enable_btf

Enables “btf”.

enable_btf_in_value

Enables “btf” in the given register value, returning the new value.

enable_btint

Enables “btint”.

enable_btint_in_value

Enables “btint” in the given register value, returning the new value.

enable_bts

Enables “bts”.

enable_bts_in_value

Enables “bts” in the given register value, returning the new value.

enable_enable_uncore_pmi

Enables “enable_uncore_pmi”.

enable_enable_uncore_pmi_in_value

Enables “enable_uncore_pmi” in the given register value, returning the new value.

enable_freeze_lbrs_on_pmi

Enables “freeze_lbrs_on_pmi”.

enable_freeze_lbrs_on_pmi_in_value

Enables “freeze_lbrs_on_pmi” in the given register value, returning the new value.

enable_freeze_perfmon_on_pmi

Enables “freeze_perfmon_on_pmi”.

enable_freeze_perfmon_on_pmi_in_value

Enables “freeze_perfmon_on_pmi” in the given register value, returning the new value.

enable_freeze_while_smm

Enables “freeze_while_smm”.

enable_freeze_while_smm_in_value

Enables “freeze_while_smm” in the given register value, returning the new value.

enable_lbr

Enables “lbr”.

enable_lbr_in_value

Enables “lbr” in the given register value, returning the new value.

enable_rtm_debug

Enables “rtm_debug”.

enable_rtm_debug_in_value

Enables “rtm_debug” in the given register value, returning the new value.

enable_tr

Enables “tr”.

enable_tr_in_value

Enables “tr” in the given register value, returning the new value.

enable_uncore_pmi_is_disabled

Returns whether “enable_uncore_pmi” is currently disabled.

enable_uncore_pmi_is_disabled_in_value

Returns whether “enable_uncore_pmi” is disabled in the given register value.

enable_uncore_pmi_is_enabled

Returns whether “enable_uncore_pmi” is currently enabled.

enable_uncore_pmi_is_enabled_in_value

Returns whether “enable_uncore_pmi” is enabled in the given register value.

freeze_lbrs_on_pmi_is_disabled

Returns whether “freeze_lbrs_on_pmi” is currently disabled.

freeze_lbrs_on_pmi_is_disabled_in_value

Returns whether “freeze_lbrs_on_pmi” is disabled in the given register value.

freeze_lbrs_on_pmi_is_enabled

Returns whether “freeze_lbrs_on_pmi” is currently enabled.

freeze_lbrs_on_pmi_is_enabled_in_value

Returns whether “freeze_lbrs_on_pmi” is enabled in the given register value.

freeze_perfmon_on_pmi_is_disabled

Returns whether “freeze_perfmon_on_pmi” is currently disabled.

freeze_perfmon_on_pmi_is_disabled_in_value

Returns whether “freeze_perfmon_on_pmi” is disabled in the given register value.

freeze_perfmon_on_pmi_is_enabled

Returns whether “freeze_perfmon_on_pmi” is currently enabled.

freeze_perfmon_on_pmi_is_enabled_in_value

Returns whether “freeze_perfmon_on_pmi” is enabled in the given register value.

freeze_while_smm_is_disabled

Returns whether “freeze_while_smm” is currently disabled.

freeze_while_smm_is_disabled_in_value

Returns whether “freeze_while_smm” is disabled in the given register value.

freeze_while_smm_is_enabled

Returns whether “freeze_while_smm” is currently enabled.

freeze_while_smm_is_enabled_in_value

Returns whether “freeze_while_smm” is enabled in the given register value.

get

Returns the current value of the register.

lbr_is_disabled

Returns whether “lbr” is currently disabled.

lbr_is_disabled_in_value

Returns whether “lbr” is disabled in the given register value.

lbr_is_enabled

Returns whether “lbr” is currently enabled.

lbr_is_enabled_in_value

Returns whether “lbr” is enabled in the given register value.

rtm_debug_is_disabled

Returns whether “rtm_debug” is currently disabled.

rtm_debug_is_disabled_in_value

Returns whether “rtm_debug” is disabled in the given register value.

rtm_debug_is_enabled

Returns whether “rtm_debug” is currently enabled.

rtm_debug_is_enabled_in_value

Returns whether “rtm_debug” is enabled in the given register value.

set

Sets the value of the register.

tr_is_disabled

Returns whether “tr” is currently disabled.

tr_is_disabled_in_value

Returns whether “tr” is disabled in the given register value.

tr_is_enabled

Returns whether “tr” is currently enabled.

tr_is_enabled_in_value

Returns whether “tr” is enabled in the given register value.