Module pal::vmcs::guest_interruptibility_state

Guest interruptibility state.

The IA-32 architecture includes features that permit certain events to be blocked for a period of time. This field contains information about such blocking. Details and the format of this field are given in Table 24-3.

Constants

ENCODING

LONG_NAME

NAME

Functions

blocking_by_mov_ss_is_disabled

Returns whether “Blocking by MOV SS” is currently disabled.

blocking_by_mov_ss_is_disabled_in_value

Returns whether “Blocking by MOV SS” is disabled in the given register value.

blocking_by_mov_ss_is_enabled

Returns whether “Blocking by MOV SS” is currently enabled.

blocking_by_mov_ss_is_enabled_in_value

Returns whether “Blocking by MOV SS” is enabled in the given register value.

blocking_by_nmi_is_disabled

Returns whether “Blocking by NMI” is currently disabled.

blocking_by_nmi_is_disabled_in_value

Returns whether “Blocking by NMI” is disabled in the given register value.

blocking_by_nmi_is_enabled

Returns whether “Blocking by NMI” is currently enabled.

blocking_by_nmi_is_enabled_in_value

Returns whether “Blocking by NMI” is enabled in the given register value.

blocking_by_smi_is_disabled

Returns whether “Blocking by SMI” is currently disabled.

blocking_by_smi_is_disabled_in_value

Returns whether “Blocking by SMI” is disabled in the given register value.

blocking_by_smi_is_enabled

Returns whether “Blocking by SMI” is currently enabled.

blocking_by_smi_is_enabled_in_value

Returns whether “Blocking by SMI” is enabled in the given register value.

blocking_by_sti_is_disabled

Returns whether “Blocking by STI” is currently disabled.

blocking_by_sti_is_disabled_in_value

Returns whether “Blocking by STI” is disabled in the given register value.

blocking_by_sti_is_enabled

Returns whether “Blocking by STI” is currently enabled.

blocking_by_sti_is_enabled_in_value

Returns whether “Blocking by STI” is enabled in the given register value.

disable_blocking_by_mov_ss

Disables “Blocking by MOV SS”.

disable_blocking_by_mov_ss_in_value

Disables “Blocking by MOV SS” in the given register value, returning the new value.

disable_blocking_by_nmi

Disables “Blocking by NMI”.

disable_blocking_by_nmi_in_value

Disables “Blocking by NMI” in the given register value, returning the new value.

disable_blocking_by_smi

Disables “Blocking by SMI”.

disable_blocking_by_smi_in_value

Disables “Blocking by SMI” in the given register value, returning the new value.

disable_blocking_by_sti

Disables “Blocking by STI”.

disable_blocking_by_sti_in_value

Disables “Blocking by STI” in the given register value, returning the new value.

disable_enclave_interruption

Disables “Enclave interruption”.

disable_enclave_interruption_in_value

Disables “Enclave interruption” in the given register value, returning the new value.

enable_blocking_by_mov_ss

Enables “Blocking by MOV SS”.

enable_blocking_by_mov_ss_in_value

Enables “Blocking by MOV SS” in the given register value, returning the new value.

enable_blocking_by_nmi

Enables “Blocking by NMI”.

enable_blocking_by_nmi_in_value

Enables “Blocking by NMI” in the given register value, returning the new value.

enable_blocking_by_smi

Enables “Blocking by SMI”.

enable_blocking_by_smi_in_value

Enables “Blocking by SMI” in the given register value, returning the new value.

enable_blocking_by_sti

Enables “Blocking by STI”.

enable_blocking_by_sti_in_value

Enables “Blocking by STI” in the given register value, returning the new value.

enable_enclave_interruption

Enables “Enclave interruption”.

enable_enclave_interruption_in_value

Enables “Enclave interruption” in the given register value, returning the new value.

enclave_interruption_is_disabled

Returns whether “Enclave interruption” is currently disabled.

enclave_interruption_is_disabled_in_value

Returns whether “Enclave interruption” is disabled in the given register value.

enclave_interruption_is_enabled

Returns whether “Enclave interruption” is currently enabled.

enclave_interruption_is_enabled_in_value

Returns whether “Enclave interruption” is enabled in the given register value.

get

Returns the current value of the register.

get_reserved

Returns the current value of “Reserved”.

get_reserved_from_value

Returns the value of “Reserved” in the given register value.

set

Sets the value of the register.

set_reserved

Sets the value of “Reserved”.

set_reserved_in_value

Sets the value of “Reserved” in the given register value, returning the new value.