Module pal::vmcs::guest_pending_debug_exceptions

Guest pending debug exceptions.

(64 bits; 32 bits on processors that do not support Intel 64 architecture). IA-32 processors may recognize one or more debug exceptions without immediately delivering them.1 This field contains information about such exceptions. This field is described in Table 24-4.

Constants

ENCODING

LONG_NAME

NAME

Functions

bs_is_disabled

Returns whether “BS” is currently disabled.

bs_is_disabled_in_value

Returns whether “BS” is disabled in the given register value.

bs_is_enabled

Returns whether “BS” is currently enabled.

bs_is_enabled_in_value

Returns whether “BS” is enabled in the given register value.

disable_bs

Disables “BS”.

disable_bs_in_value

Disables “BS” in the given register value, returning the new value.

disable_enabled_breakpoint

Disables “Enabled breakpoint”.

disable_enabled_breakpoint_in_value

Disables “Enabled breakpoint” in the given register value, returning the new value.

disable_rtm

Disables “RTM”.

disable_rtm_in_value

Disables “RTM” in the given register value, returning the new value.

enable_bs

Enables “BS”.

enable_bs_in_value

Enables “BS” in the given register value, returning the new value.

enable_enabled_breakpoint

Enables “Enabled breakpoint”.

enable_enabled_breakpoint_in_value

Enables “Enabled breakpoint” in the given register value, returning the new value.

enable_rtm

Enables “RTM”.

enable_rtm_in_value

Enables “RTM” in the given register value, returning the new value.

enabled_breakpoint_is_disabled

Returns whether “Enabled breakpoint” is currently disabled.

enabled_breakpoint_is_disabled_in_value

Returns whether “Enabled breakpoint” is disabled in the given register value.

enabled_breakpoint_is_enabled

Returns whether “Enabled breakpoint” is currently enabled.

enabled_breakpoint_is_enabled_in_value

Returns whether “Enabled breakpoint” is enabled in the given register value.

get

Returns the current value of the register.

get_b3_b0

Returns the current value of “B3 – B0”.

get_b3_b0_from_value

Returns the value of “B3 – B0” in the given register value.

rtm_is_disabled

Returns whether “RTM” is currently disabled.

rtm_is_disabled_in_value

Returns whether “RTM” is disabled in the given register value.

rtm_is_enabled

Returns whether “RTM” is currently enabled.

rtm_is_enabled_in_value

Returns whether “RTM” is enabled in the given register value.

set

Sets the value of the register.

set_b3_b0

Sets the value of “B3 – B0”.

set_b3_b0_in_value

Sets the value of “B3 – B0” in the given register value, returning the new value.