Module pal::vmcs::host_ia32_perf_global_ctrl

Host IA32_PERF_GLOBAL_CTRL (full).

Constants

ENCODING

LONG_NAME

NAME

Functions

disable_fixed_ctr0

Disables “fixed_ctr0”.

disable_fixed_ctr0_in_value

Disables “fixed_ctr0” in the given register value, returning the new value.

disable_fixed_ctr1

Disables “fixed_ctr1”.

disable_fixed_ctr1_in_value

Disables “fixed_ctr1” in the given register value, returning the new value.

disable_fixed_ctr2

Disables “fixed_ctr2”.

disable_fixed_ctr2_in_value

Disables “fixed_ctr2” in the given register value, returning the new value.

disable_pmc0

Disables “pmc0”.

disable_pmc0_in_value

Disables “pmc0” in the given register value, returning the new value.

disable_pmc1

Disables “pmc1”.

disable_pmc1_in_value

Disables “pmc1” in the given register value, returning the new value.

disable_pmc2

Disables “pmc2”.

disable_pmc2_in_value

Disables “pmc2” in the given register value, returning the new value.

disable_pmc3

Disables “pmc3”.

disable_pmc3_in_value

Disables “pmc3” in the given register value, returning the new value.

disable_pmc4

Disables “pmc4”.

disable_pmc4_in_value

Disables “pmc4” in the given register value, returning the new value.

disable_pmc5

Disables “pmc5”.

disable_pmc5_in_value

Disables “pmc5” in the given register value, returning the new value.

disable_pmc6

Disables “pmc6”.

disable_pmc6_in_value

Disables “pmc6” in the given register value, returning the new value.

disable_pmc7

Disables “pmc7”.

disable_pmc7_in_value

Disables “pmc7” in the given register value, returning the new value.

enable_fixed_ctr0

Enables “fixed_ctr0”.

enable_fixed_ctr0_in_value

Enables “fixed_ctr0” in the given register value, returning the new value.

enable_fixed_ctr1

Enables “fixed_ctr1”.

enable_fixed_ctr1_in_value

Enables “fixed_ctr1” in the given register value, returning the new value.

enable_fixed_ctr2

Enables “fixed_ctr2”.

enable_fixed_ctr2_in_value

Enables “fixed_ctr2” in the given register value, returning the new value.

enable_pmc0

Enables “pmc0”.

enable_pmc0_in_value

Enables “pmc0” in the given register value, returning the new value.

enable_pmc1

Enables “pmc1”.

enable_pmc1_in_value

Enables “pmc1” in the given register value, returning the new value.

enable_pmc2

Enables “pmc2”.

enable_pmc2_in_value

Enables “pmc2” in the given register value, returning the new value.

enable_pmc3

Enables “pmc3”.

enable_pmc3_in_value

Enables “pmc3” in the given register value, returning the new value.

enable_pmc4

Enables “pmc4”.

enable_pmc4_in_value

Enables “pmc4” in the given register value, returning the new value.

enable_pmc5

Enables “pmc5”.

enable_pmc5_in_value

Enables “pmc5” in the given register value, returning the new value.

enable_pmc6

Enables “pmc6”.

enable_pmc6_in_value

Enables “pmc6” in the given register value, returning the new value.

enable_pmc7

Enables “pmc7”.

enable_pmc7_in_value

Enables “pmc7” in the given register value, returning the new value.

fixed_ctr0_is_disabled

Returns whether “fixed_ctr0” is currently disabled.

fixed_ctr0_is_disabled_in_value

Returns whether “fixed_ctr0” is disabled in the given register value.

fixed_ctr0_is_enabled

Returns whether “fixed_ctr0” is currently enabled.

fixed_ctr0_is_enabled_in_value

Returns whether “fixed_ctr0” is enabled in the given register value.

fixed_ctr1_is_disabled

Returns whether “fixed_ctr1” is currently disabled.

fixed_ctr1_is_disabled_in_value

Returns whether “fixed_ctr1” is disabled in the given register value.

fixed_ctr1_is_enabled

Returns whether “fixed_ctr1” is currently enabled.

fixed_ctr1_is_enabled_in_value

Returns whether “fixed_ctr1” is enabled in the given register value.

fixed_ctr2_is_disabled

Returns whether “fixed_ctr2” is currently disabled.

fixed_ctr2_is_disabled_in_value

Returns whether “fixed_ctr2” is disabled in the given register value.

fixed_ctr2_is_enabled

Returns whether “fixed_ctr2” is currently enabled.

fixed_ctr2_is_enabled_in_value

Returns whether “fixed_ctr2” is enabled in the given register value.

get

Returns the current value of the register.

pmc0_is_disabled

Returns whether “pmc0” is currently disabled.

pmc0_is_disabled_in_value

Returns whether “pmc0” is disabled in the given register value.

pmc0_is_enabled

Returns whether “pmc0” is currently enabled.

pmc0_is_enabled_in_value

Returns whether “pmc0” is enabled in the given register value.

pmc1_is_disabled

Returns whether “pmc1” is currently disabled.

pmc1_is_disabled_in_value

Returns whether “pmc1” is disabled in the given register value.

pmc1_is_enabled

Returns whether “pmc1” is currently enabled.

pmc1_is_enabled_in_value

Returns whether “pmc1” is enabled in the given register value.

pmc2_is_disabled

Returns whether “pmc2” is currently disabled.

pmc2_is_disabled_in_value

Returns whether “pmc2” is disabled in the given register value.

pmc2_is_enabled

Returns whether “pmc2” is currently enabled.

pmc2_is_enabled_in_value

Returns whether “pmc2” is enabled in the given register value.

pmc3_is_disabled

Returns whether “pmc3” is currently disabled.

pmc3_is_disabled_in_value

Returns whether “pmc3” is disabled in the given register value.

pmc3_is_enabled

Returns whether “pmc3” is currently enabled.

pmc3_is_enabled_in_value

Returns whether “pmc3” is enabled in the given register value.

pmc4_is_disabled

Returns whether “pmc4” is currently disabled.

pmc4_is_disabled_in_value

Returns whether “pmc4” is disabled in the given register value.

pmc4_is_enabled

Returns whether “pmc4” is currently enabled.

pmc4_is_enabled_in_value

Returns whether “pmc4” is enabled in the given register value.

pmc5_is_disabled

Returns whether “pmc5” is currently disabled.

pmc5_is_disabled_in_value

Returns whether “pmc5” is disabled in the given register value.

pmc5_is_enabled

Returns whether “pmc5” is currently enabled.

pmc5_is_enabled_in_value

Returns whether “pmc5” is enabled in the given register value.

pmc6_is_disabled

Returns whether “pmc6” is currently disabled.

pmc6_is_disabled_in_value

Returns whether “pmc6” is disabled in the given register value.

pmc6_is_enabled

Returns whether “pmc6” is currently enabled.

pmc6_is_enabled_in_value

Returns whether “pmc6” is enabled in the given register value.

pmc7_is_disabled

Returns whether “pmc7” is currently disabled.

pmc7_is_disabled_in_value

Returns whether “pmc7” is disabled in the given register value.

pmc7_is_enabled

Returns whether “pmc7” is currently enabled.

pmc7_is_enabled_in_value

Returns whether “pmc7” is enabled in the given register value.

set

Sets the value of the register.