Module pal::vmcs::pin_based_vm_execution_controls

Pin-based VM-execution controls.

Constants

ENCODING

LONG_NAME

NAME

Functions

activate_vmx_preeemption_timer_is_disabled

Returns whether “Activate VMX-preemption timer” is currently disabled.

activate_vmx_preeemption_timer_is_disabled_in_value

Returns whether “Activate VMX-preemption timer” is disabled in the given register value.

activate_vmx_preeemption_timer_is_enabled

Returns whether “Activate VMX-preemption timer” is currently enabled.

activate_vmx_preeemption_timer_is_enabled_in_value

Returns whether “Activate VMX-preemption timer” is enabled in the given register value.

disable_activate_vmx_preeemption_timer

Disables “Activate VMX-preemption timer”.

disable_activate_vmx_preeemption_timer_in_value

Disables “Activate VMX-preemption timer” in the given register value, returning the new value.

disable_external_interrupt_exiting

Disables “External-interrupt exiting”.

disable_external_interrupt_exiting_in_value

Disables “External-interrupt exiting” in the given register value, returning the new value.

disable_nmi_exiting

Disables “NMI exiting”.

disable_nmi_exiting_in_value

Disables “NMI exiting” in the given register value, returning the new value.

disable_process_posted_interrupts

Disables “Process posted interrupts”.

disable_process_posted_interrupts_in_value

Disables “Process posted interrupts” in the given register value, returning the new value.

disable_virtual_nmis

Disables “Virtual NMIs”.

disable_virtual_nmis_in_value

Disables “Virtual NMIs” in the given register value, returning the new value.

enable_activate_vmx_preeemption_timer

Enables “Activate VMX-preemption timer”.

enable_activate_vmx_preeemption_timer_in_value

Enables “Activate VMX-preemption timer” in the given register value, returning the new value.

enable_external_interrupt_exiting

Enables “External-interrupt exiting”.

enable_external_interrupt_exiting_in_value

Enables “External-interrupt exiting” in the given register value, returning the new value.

enable_nmi_exiting

Enables “NMI exiting”.

enable_nmi_exiting_in_value

Enables “NMI exiting” in the given register value, returning the new value.

enable_process_posted_interrupts

Enables “Process posted interrupts”.

enable_process_posted_interrupts_in_value

Enables “Process posted interrupts” in the given register value, returning the new value.

enable_virtual_nmis

Enables “Virtual NMIs”.

enable_virtual_nmis_in_value

Enables “Virtual NMIs” in the given register value, returning the new value.

external_interrupt_exiting_is_disabled

Returns whether “External-interrupt exiting” is currently disabled.

external_interrupt_exiting_is_disabled_in_value

Returns whether “External-interrupt exiting” is disabled in the given register value.

external_interrupt_exiting_is_enabled

Returns whether “External-interrupt exiting” is currently enabled.

external_interrupt_exiting_is_enabled_in_value

Returns whether “External-interrupt exiting” is enabled in the given register value.

get

Returns the current value of the register.

nmi_exiting_is_disabled

Returns whether “NMI exiting” is currently disabled.

nmi_exiting_is_disabled_in_value

Returns whether “NMI exiting” is disabled in the given register value.

nmi_exiting_is_enabled

Returns whether “NMI exiting” is currently enabled.

nmi_exiting_is_enabled_in_value

Returns whether “NMI exiting” is enabled in the given register value.

process_posted_interrupts_is_disabled

Returns whether “Process posted interrupts” is currently disabled.

process_posted_interrupts_is_disabled_in_value

Returns whether “Process posted interrupts” is disabled in the given register value.

process_posted_interrupts_is_enabled

Returns whether “Process posted interrupts” is currently enabled.

process_posted_interrupts_is_enabled_in_value

Returns whether “Process posted interrupts” is enabled in the given register value.

set

Sets the value of the register.

virtual_nmis_is_disabled

Returns whether “Virtual NMIs” is currently disabled.

virtual_nmis_is_disabled_in_value

Returns whether “Virtual NMIs” is disabled in the given register value.

virtual_nmis_is_enabled

Returns whether “Virtual NMIs” is currently enabled.

virtual_nmis_is_enabled_in_value

Returns whether “Virtual NMIs” is enabled in the given register value.