Module pal::vmcs::primary_processor_based_vm_execution_controls

Primary Processor-Based VM-Execution Controls.

Constants

ENCODING

LONG_NAME

NAME

Functions

activate_secondary_controls_is_disabled

Returns whether “Activate secondary controls” is currently disabled.

activate_secondary_controls_is_disabled_in_value

Returns whether “Activate secondary controls” is disabled in the given register value.

activate_secondary_controls_is_enabled

Returns whether “Activate secondary controls” is currently enabled.

activate_secondary_controls_is_enabled_in_value

Returns whether “Activate secondary controls” is enabled in the given register value.

cr3_load_exiting_is_disabled

Returns whether “CR3-load exiting” is currently disabled.

cr3_load_exiting_is_disabled_in_value

Returns whether “CR3-load exiting” is disabled in the given register value.

cr3_load_exiting_is_enabled

Returns whether “CR3-load exiting” is currently enabled.

cr3_load_exiting_is_enabled_in_value

Returns whether “CR3-load exiting” is enabled in the given register value.

cr3_store_exiting_is_disabled

Returns whether “CR3-store exiting” is currently disabled.

cr3_store_exiting_is_disabled_in_value

Returns whether “CR3-store exiting” is disabled in the given register value.

cr3_store_exiting_is_enabled

Returns whether “CR3-store exiting” is currently enabled.

cr3_store_exiting_is_enabled_in_value

Returns whether “CR3-store exiting” is enabled in the given register value.

cr8_load_exiting_is_disabled

Returns whether “CR8-load exiting” is currently disabled.

cr8_load_exiting_is_disabled_in_value

Returns whether “CR8-load exiting” is disabled in the given register value.

cr8_load_exiting_is_enabled

Returns whether “CR8-load exiting” is currently enabled.

cr8_load_exiting_is_enabled_in_value

Returns whether “CR8-load exiting” is enabled in the given register value.

cr8_store_exiting_is_disabled

Returns whether “CR8-store exiting” is currently disabled.

cr8_store_exiting_is_disabled_in_value

Returns whether “CR8-store exiting” is disabled in the given register value.

cr8_store_exiting_is_enabled

Returns whether “CR8-store exiting” is currently enabled.

cr8_store_exiting_is_enabled_in_value

Returns whether “CR8-store exiting” is enabled in the given register value.

disable_activate_secondary_controls

Disables “Activate secondary controls”.

disable_activate_secondary_controls_in_value

Disables “Activate secondary controls” in the given register value, returning the new value.

disable_cr3_load_exiting

Disables “CR3-load exiting”.

disable_cr3_load_exiting_in_value

Disables “CR3-load exiting” in the given register value, returning the new value.

disable_cr3_store_exiting

Disables “CR3-store exiting”.

disable_cr3_store_exiting_in_value

Disables “CR3-store exiting” in the given register value, returning the new value.

disable_cr8_load_exiting

Disables “CR8-load exiting”.

disable_cr8_load_exiting_in_value

Disables “CR8-load exiting” in the given register value, returning the new value.

disable_cr8_store_exiting

Disables “CR8-store exiting”.

disable_cr8_store_exiting_in_value

Disables “CR8-store exiting” in the given register value, returning the new value.

disable_hlt_exiting

Disables “HLT exiting”.

disable_hlt_exiting_in_value

Disables “HLT exiting” in the given register value, returning the new value.

disable_interrupt_window_exiting

Disables “Interrupt-window exiting”.

disable_interrupt_window_exiting_in_value

Disables “Interrupt-window exiting” in the given register value, returning the new value.

disable_invlpg_exiting

Disables “INVLPG exiting”.

disable_invlpg_exiting_in_value

Disables “INVLPG exiting” in the given register value, returning the new value.

disable_monitor_exiting

Disables “MONITOR exiting”.

disable_monitor_exiting_in_value

Disables “MONITOR exiting” in the given register value, returning the new value.

disable_monitor_trap_flag

Disables “Monitor trap flag”.

disable_monitor_trap_flag_in_value

Disables “Monitor trap flag” in the given register value, returning the new value.

disable_mov_dr_exiting

Disables “MOV-DR exiting”.

disable_mov_dr_exiting_in_value

Disables “MOV-DR exiting” in the given register value, returning the new value.

disable_mwait_exiting

Disables “MWAIT exiting”.

disable_mwait_exiting_in_value

Disables “MWAIT exiting” in the given register value, returning the new value.

disable_nmi_window_exiting

Disables “NMI-window exiting”.

disable_nmi_window_exiting_in_value

Disables “NMI-window exiting” in the given register value, returning the new value.

disable_pause_exiting

Disables “PAUSE exiting”.

disable_pause_exiting_in_value

Disables “PAUSE exiting” in the given register value, returning the new value.

disable_rdpmc_exiting

Disables “RDPMC exiting”.

disable_rdpmc_exiting_in_value

Disables “RDPMC exiting” in the given register value, returning the new value.

disable_rdtsc_exiting

Disables “RDTSC exiting”.

disable_rdtsc_exiting_in_value

Disables “RDTSC exiting” in the given register value, returning the new value.

disable_unconditional_io_exiting

Disables “ Unconditional I/O exiting“.

disable_unconditional_io_exiting_in_value

Disables “ Unconditional I/O exiting“ in the given register value, returning the new value.

disable_use_io_bitmaps

Disables “Use I/O bitmaps”.

disable_use_io_bitmaps_in_value

Disables “Use I/O bitmaps” in the given register value, returning the new value.

disable_use_msr_bitmaps

Disables “Use MSR bitmaps”.

disable_use_msr_bitmaps_in_value

Disables “Use MSR bitmaps” in the given register value, returning the new value.

disable_use_tpr_shadow

Disables “Use TPR shadow”.

disable_use_tpr_shadow_in_value

Disables “Use TPR shadow” in the given register value, returning the new value.

disable_use_tsc_offsetting

Disables “Use TSC offsetting”.

disable_use_tsc_offsetting_in_value

Disables “Use TSC offsetting” in the given register value, returning the new value.

enable_activate_secondary_controls

Enables “Activate secondary controls”.

enable_activate_secondary_controls_in_value

Enables “Activate secondary controls” in the given register value, returning the new value.

enable_cr3_load_exiting

Enables “CR3-load exiting”.

enable_cr3_load_exiting_in_value

Enables “CR3-load exiting” in the given register value, returning the new value.

enable_cr3_store_exiting

Enables “CR3-store exiting”.

enable_cr3_store_exiting_in_value

Enables “CR3-store exiting” in the given register value, returning the new value.

enable_cr8_load_exiting

Enables “CR8-load exiting”.

enable_cr8_load_exiting_in_value

Enables “CR8-load exiting” in the given register value, returning the new value.

enable_cr8_store_exiting

Enables “CR8-store exiting”.

enable_cr8_store_exiting_in_value

Enables “CR8-store exiting” in the given register value, returning the new value.

enable_hlt_exiting

Enables “HLT exiting”.

enable_hlt_exiting_in_value

Enables “HLT exiting” in the given register value, returning the new value.

enable_interrupt_window_exiting

Enables “Interrupt-window exiting”.

enable_interrupt_window_exiting_in_value

Enables “Interrupt-window exiting” in the given register value, returning the new value.

enable_invlpg_exiting

Enables “INVLPG exiting”.

enable_invlpg_exiting_in_value

Enables “INVLPG exiting” in the given register value, returning the new value.

enable_monitor_exiting

Enables “MONITOR exiting”.

enable_monitor_exiting_in_value

Enables “MONITOR exiting” in the given register value, returning the new value.

enable_monitor_trap_flag

Enables “Monitor trap flag”.

enable_monitor_trap_flag_in_value

Enables “Monitor trap flag” in the given register value, returning the new value.

enable_mov_dr_exiting

Enables “MOV-DR exiting”.

enable_mov_dr_exiting_in_value

Enables “MOV-DR exiting” in the given register value, returning the new value.

enable_mwait_exiting

Enables “MWAIT exiting”.

enable_mwait_exiting_in_value

Enables “MWAIT exiting” in the given register value, returning the new value.

enable_nmi_window_exiting

Enables “NMI-window exiting”.

enable_nmi_window_exiting_in_value

Enables “NMI-window exiting” in the given register value, returning the new value.

enable_pause_exiting

Enables “PAUSE exiting”.

enable_pause_exiting_in_value

Enables “PAUSE exiting” in the given register value, returning the new value.

enable_rdpmc_exiting

Enables “RDPMC exiting”.

enable_rdpmc_exiting_in_value

Enables “RDPMC exiting” in the given register value, returning the new value.

enable_rdtsc_exiting

Enables “RDTSC exiting”.

enable_rdtsc_exiting_in_value

Enables “RDTSC exiting” in the given register value, returning the new value.

enable_unconditional_io_exiting

Enables “ Unconditional I/O exiting“.

enable_unconditional_io_exiting_in_value

Enables “ Unconditional I/O exiting“ in the given register value, returning the new value.

enable_use_io_bitmaps

Enables “Use I/O bitmaps”.

enable_use_io_bitmaps_in_value

Enables “Use I/O bitmaps” in the given register value, returning the new value.

enable_use_msr_bitmaps

Enables “Use MSR bitmaps”.

enable_use_msr_bitmaps_in_value

Enables “Use MSR bitmaps” in the given register value, returning the new value.

enable_use_tpr_shadow

Enables “Use TPR shadow”.

enable_use_tpr_shadow_in_value

Enables “Use TPR shadow” in the given register value, returning the new value.

enable_use_tsc_offsetting

Enables “Use TSC offsetting”.

enable_use_tsc_offsetting_in_value

Enables “Use TSC offsetting” in the given register value, returning the new value.

get

Returns the current value of the register.

hlt_exiting_is_disabled

Returns whether “HLT exiting” is currently disabled.

hlt_exiting_is_disabled_in_value

Returns whether “HLT exiting” is disabled in the given register value.

hlt_exiting_is_enabled

Returns whether “HLT exiting” is currently enabled.

hlt_exiting_is_enabled_in_value

Returns whether “HLT exiting” is enabled in the given register value.

interrupt_window_exiting_is_disabled

Returns whether “Interrupt-window exiting” is currently disabled.

interrupt_window_exiting_is_disabled_in_value

Returns whether “Interrupt-window exiting” is disabled in the given register value.

interrupt_window_exiting_is_enabled

Returns whether “Interrupt-window exiting” is currently enabled.

interrupt_window_exiting_is_enabled_in_value

Returns whether “Interrupt-window exiting” is enabled in the given register value.

invlpg_exiting_is_disabled

Returns whether “INVLPG exiting” is currently disabled.

invlpg_exiting_is_disabled_in_value

Returns whether “INVLPG exiting” is disabled in the given register value.

invlpg_exiting_is_enabled

Returns whether “INVLPG exiting” is currently enabled.

invlpg_exiting_is_enabled_in_value

Returns whether “INVLPG exiting” is enabled in the given register value.

monitor_exiting_is_disabled

Returns whether “MONITOR exiting” is currently disabled.

monitor_exiting_is_disabled_in_value

Returns whether “MONITOR exiting” is disabled in the given register value.

monitor_exiting_is_enabled

Returns whether “MONITOR exiting” is currently enabled.

monitor_exiting_is_enabled_in_value

Returns whether “MONITOR exiting” is enabled in the given register value.

monitor_trap_flag_is_disabled

Returns whether “Monitor trap flag” is currently disabled.

monitor_trap_flag_is_disabled_in_value

Returns whether “Monitor trap flag” is disabled in the given register value.

monitor_trap_flag_is_enabled

Returns whether “Monitor trap flag” is currently enabled.

monitor_trap_flag_is_enabled_in_value

Returns whether “Monitor trap flag” is enabled in the given register value.

mov_dr_exiting_is_disabled

Returns whether “MOV-DR exiting” is currently disabled.

mov_dr_exiting_is_disabled_in_value

Returns whether “MOV-DR exiting” is disabled in the given register value.

mov_dr_exiting_is_enabled

Returns whether “MOV-DR exiting” is currently enabled.

mov_dr_exiting_is_enabled_in_value

Returns whether “MOV-DR exiting” is enabled in the given register value.

mwait_exiting_is_disabled

Returns whether “MWAIT exiting” is currently disabled.

mwait_exiting_is_disabled_in_value

Returns whether “MWAIT exiting” is disabled in the given register value.

mwait_exiting_is_enabled

Returns whether “MWAIT exiting” is currently enabled.

mwait_exiting_is_enabled_in_value

Returns whether “MWAIT exiting” is enabled in the given register value.

nmi_window_exiting_is_disabled

Returns whether “NMI-window exiting” is currently disabled.

nmi_window_exiting_is_disabled_in_value

Returns whether “NMI-window exiting” is disabled in the given register value.

nmi_window_exiting_is_enabled

Returns whether “NMI-window exiting” is currently enabled.

nmi_window_exiting_is_enabled_in_value

Returns whether “NMI-window exiting” is enabled in the given register value.

pause_exiting_is_disabled

Returns whether “PAUSE exiting” is currently disabled.

pause_exiting_is_disabled_in_value

Returns whether “PAUSE exiting” is disabled in the given register value.

pause_exiting_is_enabled

Returns whether “PAUSE exiting” is currently enabled.

pause_exiting_is_enabled_in_value

Returns whether “PAUSE exiting” is enabled in the given register value.

rdpmc_exiting_is_disabled

Returns whether “RDPMC exiting” is currently disabled.

rdpmc_exiting_is_disabled_in_value

Returns whether “RDPMC exiting” is disabled in the given register value.

rdpmc_exiting_is_enabled

Returns whether “RDPMC exiting” is currently enabled.

rdpmc_exiting_is_enabled_in_value

Returns whether “RDPMC exiting” is enabled in the given register value.

rdtsc_exiting_is_disabled

Returns whether “RDTSC exiting” is currently disabled.

rdtsc_exiting_is_disabled_in_value

Returns whether “RDTSC exiting” is disabled in the given register value.

rdtsc_exiting_is_enabled

Returns whether “RDTSC exiting” is currently enabled.

rdtsc_exiting_is_enabled_in_value

Returns whether “RDTSC exiting” is enabled in the given register value.

set

Sets the value of the register.

unconditional_io_exiting_is_disabled

Returns whether “ Unconditional I/O exiting“ is currently disabled.

unconditional_io_exiting_is_disabled_in_value

Returns whether “ Unconditional I/O exiting“ is disabled in the given register value.

unconditional_io_exiting_is_enabled

Returns whether “ Unconditional I/O exiting“ is currently enabled.

unconditional_io_exiting_is_enabled_in_value

Returns whether “ Unconditional I/O exiting“ is enabled in the given register value.

use_io_bitmaps_is_disabled

Returns whether “Use I/O bitmaps” is currently disabled.

use_io_bitmaps_is_disabled_in_value

Returns whether “Use I/O bitmaps” is disabled in the given register value.

use_io_bitmaps_is_enabled

Returns whether “Use I/O bitmaps” is currently enabled.

use_io_bitmaps_is_enabled_in_value

Returns whether “Use I/O bitmaps” is enabled in the given register value.

use_msr_bitmaps_is_disabled

Returns whether “Use MSR bitmaps” is currently disabled.

use_msr_bitmaps_is_disabled_in_value

Returns whether “Use MSR bitmaps” is disabled in the given register value.

use_msr_bitmaps_is_enabled

Returns whether “Use MSR bitmaps” is currently enabled.

use_msr_bitmaps_is_enabled_in_value

Returns whether “Use MSR bitmaps” is enabled in the given register value.

use_tpr_shadow_is_disabled

Returns whether “Use TPR shadow” is currently disabled.

use_tpr_shadow_is_disabled_in_value

Returns whether “Use TPR shadow” is disabled in the given register value.

use_tpr_shadow_is_enabled

Returns whether “Use TPR shadow” is currently enabled.

use_tpr_shadow_is_enabled_in_value

Returns whether “Use TPR shadow” is enabled in the given register value.

use_tsc_offsetting_is_disabled

Returns whether “Use TSC offsetting” is currently disabled.

use_tsc_offsetting_is_disabled_in_value

Returns whether “Use TSC offsetting” is disabled in the given register value.

use_tsc_offsetting_is_enabled

Returns whether “Use TSC offsetting” is currently enabled.

use_tsc_offsetting_is_enabled_in_value

Returns whether “Use TSC offsetting” is enabled in the given register value.