Module pal::vmcs::secondary_processor_based_vm_execution_controls

Secondary Processor-Based VM-Execution Controls.

Constants

ENCODING

LONG_NAME

NAME

Functions

apic_register_virtualization_is_disabled

Returns whether “ APIC-register virtualization“ is currently disabled.

apic_register_virtualization_is_disabled_in_value

Returns whether “ APIC-register virtualization“ is disabled in the given register value.

apic_register_virtualization_is_enabled

Returns whether “ APIC-register virtualization“ is currently enabled.

apic_register_virtualization_is_enabled_in_value

Returns whether “ APIC-register virtualization“ is enabled in the given register value.

conceal_vmx_from_pt_is_disabled

Returns whether “Conceal VMX from PT” is currently disabled.

conceal_vmx_from_pt_is_disabled_in_value

Returns whether “Conceal VMX from PT” is disabled in the given register value.

conceal_vmx_from_pt_is_enabled

Returns whether “Conceal VMX from PT” is currently enabled.

conceal_vmx_from_pt_is_enabled_in_value

Returns whether “Conceal VMX from PT” is enabled in the given register value.

descriptor_table_exiting_is_disabled

Returns whether “Descriptor-table exiting” is currently disabled.

descriptor_table_exiting_is_disabled_in_value

Returns whether “Descriptor-table exiting” is disabled in the given register value.

descriptor_table_exiting_is_enabled

Returns whether “Descriptor-table exiting” is currently enabled.

descriptor_table_exiting_is_enabled_in_value

Returns whether “Descriptor-table exiting” is enabled in the given register value.

disable_apic_register_virtualization

Disables “ APIC-register virtualization“.

disable_apic_register_virtualization_in_value

Disables “ APIC-register virtualization“ in the given register value, returning the new value.

disable_conceal_vmx_from_pt

Disables “Conceal VMX from PT”.

disable_conceal_vmx_from_pt_in_value

Disables “Conceal VMX from PT” in the given register value, returning the new value.

disable_descriptor_table_exiting

Disables “Descriptor-table exiting”.

disable_descriptor_table_exiting_in_value

Disables “Descriptor-table exiting” in the given register value, returning the new value.

disable_enable_encls_exiting

Disables “Enable ENCLS exiting”.

disable_enable_encls_exiting_in_value

Disables “Enable ENCLS exiting” in the given register value, returning the new value.

disable_enable_enclv_exiting

Disables “Enable ENCLV exiting”.

disable_enable_enclv_exiting_in_value

Disables “Enable ENCLV exiting” in the given register value, returning the new value.

disable_enable_ept

Disables “Enable EPT”.

disable_enable_ept_in_value

Disables “Enable EPT” in the given register value, returning the new value.

disable_enable_invpcid

Disables “Enable INVPCID”.

disable_enable_invpcid_in_value

Disables “Enable INVPCID” in the given register value, returning the new value.

disable_enable_pml

Disables “Enable PML”.

disable_enable_pml_in_value

Disables “Enable PML” in the given register value, returning the new value.

disable_enable_rdtscp

Disables “Enable RDTSCP”.

disable_enable_rdtscp_in_value

Disables “Enable RDTSCP” in the given register value, returning the new value.

disable_enable_use_wait_and_pause

Disables “ Enable user wait and pause“.

disable_enable_use_wait_and_pause_in_value

Disables “ Enable user wait and pause“ in the given register value, returning the new value.

disable_enable_vm_functions

Disables “ Enable VM functions“.

disable_enable_vm_functions_in_value

Disables “ Enable VM functions“ in the given register value, returning the new value.

disable_enable_vpid

Disables “Enable VPID”.

disable_enable_vpid_in_value

Disables “Enable VPID” in the given register value, returning the new value.

disable_enable_xsaves_xrstors

Disables “ Enable XSAVES/XRSTORS“.

disable_enable_xsaves_xrstors_in_value

Disables “ Enable XSAVES/XRSTORS“ in the given register value, returning the new value.

disable_ept_violation_ve

Disables “EPT-violation #VE”.

disable_ept_violation_ve_in_value

Disables “EPT-violation #VE” in the given register value, returning the new value.

disable_intel_pt_uses_guest_physical_addresses

Disables “Intel PT uses guest physical addresses”.

disable_intel_pt_uses_guest_physical_addresses_in_value

Disables “Intel PT uses guest physical addresses” in the given register value, returning the new value.

disable_mode_based_execute_control_for_ept

Disables “Mode-based execute control for EPT”.

disable_mode_based_execute_control_for_ept_in_value

Disables “Mode-based execute control for EPT” in the given register value, returning the new value.

disable_pause_loop_exiting

Disables “PAUSE-loop exiting”.

disable_pause_loop_exiting_in_value

Disables “PAUSE-loop exiting” in the given register value, returning the new value.

disable_rdrand_exiting

Disables “RDRAND exiting”.

disable_rdrand_exiting_in_value

Disables “RDRAND exiting” in the given register value, returning the new value.

disable_rdseed_exiting

Disables “RDSEED exiting”.

disable_rdseed_exiting_in_value

Disables “RDSEED exiting” in the given register value, returning the new value.

disable_sub_page_write_permissions_for_ept

Disables “ Sub-page write permissions for EPT“.

disable_sub_page_write_permissions_for_ept_in_value

Disables “ Sub-page write permissions for EPT“ in the given register value, returning the new value.

disable_urestricted_guest

Disables “Unrestricted guest”.

disable_urestricted_guest_in_value

Disables “Unrestricted guest” in the given register value, returning the new value.

disable_use_tsc_scaling

Disables “Use TSC scaling”.

disable_use_tsc_scaling_in_value

Disables “Use TSC scaling” in the given register value, returning the new value.

disable_virtual_interrupt_delivery

Disables “Virtual-interrupt delivery”.

disable_virtual_interrupt_delivery_in_value

Disables “Virtual-interrupt delivery” in the given register value, returning the new value.

disable_virtualize_apic_accesses

Disables “Virtualize APIC accesses”.

disable_virtualize_apic_accesses_in_value

Disables “Virtualize APIC accesses” in the given register value, returning the new value.

disable_virtualize_x2apic_mode

Disables “Virtualize x2APIC mode”.

disable_virtualize_x2apic_mode_in_value

Disables “Virtualize x2APIC mode” in the given register value, returning the new value.

disable_vmcs_shadowing

Disables “VMCS shadowing”.

disable_vmcs_shadowing_in_value

Disables “VMCS shadowing” in the given register value, returning the new value.

disable_wbinvd_exiting

Disables “WBINVD exiting”.

disable_wbinvd_exiting_in_value

Disables “WBINVD exiting” in the given register value, returning the new value.

enable_apic_register_virtualization

Enables “ APIC-register virtualization“.

enable_apic_register_virtualization_in_value

Enables “ APIC-register virtualization“ in the given register value, returning the new value.

enable_conceal_vmx_from_pt

Enables “Conceal VMX from PT”.

enable_conceal_vmx_from_pt_in_value

Enables “Conceal VMX from PT” in the given register value, returning the new value.

enable_descriptor_table_exiting

Enables “Descriptor-table exiting”.

enable_descriptor_table_exiting_in_value

Enables “Descriptor-table exiting” in the given register value, returning the new value.

enable_enable_encls_exiting

Enables “Enable ENCLS exiting”.

enable_enable_encls_exiting_in_value

Enables “Enable ENCLS exiting” in the given register value, returning the new value.

enable_enable_enclv_exiting

Enables “Enable ENCLV exiting”.

enable_enable_enclv_exiting_in_value

Enables “Enable ENCLV exiting” in the given register value, returning the new value.

enable_enable_ept

Enables “Enable EPT”.

enable_enable_ept_in_value

Enables “Enable EPT” in the given register value, returning the new value.

enable_enable_invpcid

Enables “Enable INVPCID”.

enable_enable_invpcid_in_value

Enables “Enable INVPCID” in the given register value, returning the new value.

enable_enable_pml

Enables “Enable PML”.

enable_enable_pml_in_value

Enables “Enable PML” in the given register value, returning the new value.

enable_enable_rdtscp

Enables “Enable RDTSCP”.

enable_enable_rdtscp_in_value

Enables “Enable RDTSCP” in the given register value, returning the new value.

enable_enable_use_wait_and_pause

Enables “ Enable user wait and pause“.

enable_enable_use_wait_and_pause_in_value

Enables “ Enable user wait and pause“ in the given register value, returning the new value.

enable_enable_vm_functions

Enables “ Enable VM functions“.

enable_enable_vm_functions_in_value

Enables “ Enable VM functions“ in the given register value, returning the new value.

enable_enable_vpid

Enables “Enable VPID”.

enable_enable_vpid_in_value

Enables “Enable VPID” in the given register value, returning the new value.

enable_enable_xsaves_xrstors

Enables “ Enable XSAVES/XRSTORS“.

enable_enable_xsaves_xrstors_in_value

Enables “ Enable XSAVES/XRSTORS“ in the given register value, returning the new value.

enable_encls_exiting_is_disabled

Returns whether “Enable ENCLS exiting” is currently disabled.

enable_encls_exiting_is_disabled_in_value

Returns whether “Enable ENCLS exiting” is disabled in the given register value.

enable_encls_exiting_is_enabled

Returns whether “Enable ENCLS exiting” is currently enabled.

enable_encls_exiting_is_enabled_in_value

Returns whether “Enable ENCLS exiting” is enabled in the given register value.

enable_enclv_exiting_is_disabled

Returns whether “Enable ENCLV exiting” is currently disabled.

enable_enclv_exiting_is_disabled_in_value

Returns whether “Enable ENCLV exiting” is disabled in the given register value.

enable_enclv_exiting_is_enabled

Returns whether “Enable ENCLV exiting” is currently enabled.

enable_enclv_exiting_is_enabled_in_value

Returns whether “Enable ENCLV exiting” is enabled in the given register value.

enable_ept_is_disabled

Returns whether “Enable EPT” is currently disabled.

enable_ept_is_disabled_in_value

Returns whether “Enable EPT” is disabled in the given register value.

enable_ept_is_enabled

Returns whether “Enable EPT” is currently enabled.

enable_ept_is_enabled_in_value

Returns whether “Enable EPT” is enabled in the given register value.

enable_ept_violation_ve

Enables “EPT-violation #VE”.

enable_ept_violation_ve_in_value

Enables “EPT-violation #VE” in the given register value, returning the new value.

enable_intel_pt_uses_guest_physical_addresses

Enables “Intel PT uses guest physical addresses”.

enable_intel_pt_uses_guest_physical_addresses_in_value

Enables “Intel PT uses guest physical addresses” in the given register value, returning the new value.

enable_invpcid_is_disabled

Returns whether “Enable INVPCID” is currently disabled.

enable_invpcid_is_disabled_in_value

Returns whether “Enable INVPCID” is disabled in the given register value.

enable_invpcid_is_enabled

Returns whether “Enable INVPCID” is currently enabled.

enable_invpcid_is_enabled_in_value

Returns whether “Enable INVPCID” is enabled in the given register value.

enable_mode_based_execute_control_for_ept

Enables “Mode-based execute control for EPT”.

enable_mode_based_execute_control_for_ept_in_value

Enables “Mode-based execute control for EPT” in the given register value, returning the new value.

enable_pause_loop_exiting

Enables “PAUSE-loop exiting”.

enable_pause_loop_exiting_in_value

Enables “PAUSE-loop exiting” in the given register value, returning the new value.

enable_pml_is_disabled

Returns whether “Enable PML” is currently disabled.

enable_pml_is_disabled_in_value

Returns whether “Enable PML” is disabled in the given register value.

enable_pml_is_enabled

Returns whether “Enable PML” is currently enabled.

enable_pml_is_enabled_in_value

Returns whether “Enable PML” is enabled in the given register value.

enable_rdrand_exiting

Enables “RDRAND exiting”.

enable_rdrand_exiting_in_value

Enables “RDRAND exiting” in the given register value, returning the new value.

enable_rdseed_exiting

Enables “RDSEED exiting”.

enable_rdseed_exiting_in_value

Enables “RDSEED exiting” in the given register value, returning the new value.

enable_rdtscp_is_disabled

Returns whether “Enable RDTSCP” is currently disabled.

enable_rdtscp_is_disabled_in_value

Returns whether “Enable RDTSCP” is disabled in the given register value.

enable_rdtscp_is_enabled

Returns whether “Enable RDTSCP” is currently enabled.

enable_rdtscp_is_enabled_in_value

Returns whether “Enable RDTSCP” is enabled in the given register value.

enable_sub_page_write_permissions_for_ept

Enables “ Sub-page write permissions for EPT“.

enable_sub_page_write_permissions_for_ept_in_value

Enables “ Sub-page write permissions for EPT“ in the given register value, returning the new value.

enable_urestricted_guest

Enables “Unrestricted guest”.

enable_urestricted_guest_in_value

Enables “Unrestricted guest” in the given register value, returning the new value.

enable_use_tsc_scaling

Enables “Use TSC scaling”.

enable_use_tsc_scaling_in_value

Enables “Use TSC scaling” in the given register value, returning the new value.

enable_use_wait_and_pause_is_disabled

Returns whether “ Enable user wait and pause“ is currently disabled.

enable_use_wait_and_pause_is_disabled_in_value

Returns whether “ Enable user wait and pause“ is disabled in the given register value.

enable_use_wait_and_pause_is_enabled

Returns whether “ Enable user wait and pause“ is currently enabled.

enable_use_wait_and_pause_is_enabled_in_value

Returns whether “ Enable user wait and pause“ is enabled in the given register value.

enable_virtual_interrupt_delivery

Enables “Virtual-interrupt delivery”.

enable_virtual_interrupt_delivery_in_value

Enables “Virtual-interrupt delivery” in the given register value, returning the new value.

enable_virtualize_apic_accesses

Enables “Virtualize APIC accesses”.

enable_virtualize_apic_accesses_in_value

Enables “Virtualize APIC accesses” in the given register value, returning the new value.

enable_virtualize_x2apic_mode

Enables “Virtualize x2APIC mode”.

enable_virtualize_x2apic_mode_in_value

Enables “Virtualize x2APIC mode” in the given register value, returning the new value.

enable_vm_functions_is_disabled

Returns whether “ Enable VM functions“ is currently disabled.

enable_vm_functions_is_disabled_in_value

Returns whether “ Enable VM functions“ is disabled in the given register value.

enable_vm_functions_is_enabled

Returns whether “ Enable VM functions“ is currently enabled.

enable_vm_functions_is_enabled_in_value

Returns whether “ Enable VM functions“ is enabled in the given register value.

enable_vmcs_shadowing

Enables “VMCS shadowing”.

enable_vmcs_shadowing_in_value

Enables “VMCS shadowing” in the given register value, returning the new value.

enable_vpid_is_disabled

Returns whether “Enable VPID” is currently disabled.

enable_vpid_is_disabled_in_value

Returns whether “Enable VPID” is disabled in the given register value.

enable_vpid_is_enabled

Returns whether “Enable VPID” is currently enabled.

enable_vpid_is_enabled_in_value

Returns whether “Enable VPID” is enabled in the given register value.

enable_wbinvd_exiting

Enables “WBINVD exiting”.

enable_wbinvd_exiting_in_value

Enables “WBINVD exiting” in the given register value, returning the new value.

enable_xsaves_xrstors_is_disabled

Returns whether “ Enable XSAVES/XRSTORS“ is currently disabled.

enable_xsaves_xrstors_is_disabled_in_value

Returns whether “ Enable XSAVES/XRSTORS“ is disabled in the given register value.

enable_xsaves_xrstors_is_enabled

Returns whether “ Enable XSAVES/XRSTORS“ is currently enabled.

enable_xsaves_xrstors_is_enabled_in_value

Returns whether “ Enable XSAVES/XRSTORS“ is enabled in the given register value.

ept_violation_ve_is_disabled

Returns whether “EPT-violation #VE” is currently disabled.

ept_violation_ve_is_disabled_in_value

Returns whether “EPT-violation #VE” is disabled in the given register value.

ept_violation_ve_is_enabled

Returns whether “EPT-violation #VE” is currently enabled.

ept_violation_ve_is_enabled_in_value

Returns whether “EPT-violation #VE” is enabled in the given register value.

get

Returns the current value of the register.

intel_pt_uses_guest_physical_addresses_is_disabled

Returns whether “Intel PT uses guest physical addresses” is currently disabled.

intel_pt_uses_guest_physical_addresses_is_disabled_in_value

Returns whether “Intel PT uses guest physical addresses” is disabled in the given register value.

intel_pt_uses_guest_physical_addresses_is_enabled

Returns whether “Intel PT uses guest physical addresses” is currently enabled.

intel_pt_uses_guest_physical_addresses_is_enabled_in_value

Returns whether “Intel PT uses guest physical addresses” is enabled in the given register value.

mode_based_execute_control_for_ept_is_disabled

Returns whether “Mode-based execute control for EPT” is currently disabled.

mode_based_execute_control_for_ept_is_disabled_in_value

Returns whether “Mode-based execute control for EPT” is disabled in the given register value.

mode_based_execute_control_for_ept_is_enabled

Returns whether “Mode-based execute control for EPT” is currently enabled.

mode_based_execute_control_for_ept_is_enabled_in_value

Returns whether “Mode-based execute control for EPT” is enabled in the given register value.

pause_loop_exiting_is_disabled

Returns whether “PAUSE-loop exiting” is currently disabled.

pause_loop_exiting_is_disabled_in_value

Returns whether “PAUSE-loop exiting” is disabled in the given register value.

pause_loop_exiting_is_enabled

Returns whether “PAUSE-loop exiting” is currently enabled.

pause_loop_exiting_is_enabled_in_value

Returns whether “PAUSE-loop exiting” is enabled in the given register value.

rdrand_exiting_is_disabled

Returns whether “RDRAND exiting” is currently disabled.

rdrand_exiting_is_disabled_in_value

Returns whether “RDRAND exiting” is disabled in the given register value.

rdrand_exiting_is_enabled

Returns whether “RDRAND exiting” is currently enabled.

rdrand_exiting_is_enabled_in_value

Returns whether “RDRAND exiting” is enabled in the given register value.

rdseed_exiting_is_disabled

Returns whether “RDSEED exiting” is currently disabled.

rdseed_exiting_is_disabled_in_value

Returns whether “RDSEED exiting” is disabled in the given register value.

rdseed_exiting_is_enabled

Returns whether “RDSEED exiting” is currently enabled.

rdseed_exiting_is_enabled_in_value

Returns whether “RDSEED exiting” is enabled in the given register value.

set

Sets the value of the register.

sub_page_write_permissions_for_ept_is_disabled

Returns whether “ Sub-page write permissions for EPT“ is currently disabled.

sub_page_write_permissions_for_ept_is_disabled_in_value

Returns whether “ Sub-page write permissions for EPT“ is disabled in the given register value.

sub_page_write_permissions_for_ept_is_enabled

Returns whether “ Sub-page write permissions for EPT“ is currently enabled.

sub_page_write_permissions_for_ept_is_enabled_in_value

Returns whether “ Sub-page write permissions for EPT“ is enabled in the given register value.

urestricted_guest_is_disabled

Returns whether “Unrestricted guest” is currently disabled.

urestricted_guest_is_disabled_in_value

Returns whether “Unrestricted guest” is disabled in the given register value.

urestricted_guest_is_enabled

Returns whether “Unrestricted guest” is currently enabled.

urestricted_guest_is_enabled_in_value

Returns whether “Unrestricted guest” is enabled in the given register value.

use_tsc_scaling_is_disabled

Returns whether “Use TSC scaling” is currently disabled.

use_tsc_scaling_is_disabled_in_value

Returns whether “Use TSC scaling” is disabled in the given register value.

use_tsc_scaling_is_enabled

Returns whether “Use TSC scaling” is currently enabled.

use_tsc_scaling_is_enabled_in_value

Returns whether “Use TSC scaling” is enabled in the given register value.

virtual_interrupt_delivery_is_disabled

Returns whether “Virtual-interrupt delivery” is currently disabled.

virtual_interrupt_delivery_is_disabled_in_value

Returns whether “Virtual-interrupt delivery” is disabled in the given register value.

virtual_interrupt_delivery_is_enabled

Returns whether “Virtual-interrupt delivery” is currently enabled.

virtual_interrupt_delivery_is_enabled_in_value

Returns whether “Virtual-interrupt delivery” is enabled in the given register value.

virtualize_apic_accesses_is_disabled

Returns whether “Virtualize APIC accesses” is currently disabled.

virtualize_apic_accesses_is_disabled_in_value

Returns whether “Virtualize APIC accesses” is disabled in the given register value.

virtualize_apic_accesses_is_enabled

Returns whether “Virtualize APIC accesses” is currently enabled.

virtualize_apic_accesses_is_enabled_in_value

Returns whether “Virtualize APIC accesses” is enabled in the given register value.

virtualize_x2apic_mode_is_disabled

Returns whether “Virtualize x2APIC mode” is currently disabled.

virtualize_x2apic_mode_is_disabled_in_value

Returns whether “Virtualize x2APIC mode” is disabled in the given register value.

virtualize_x2apic_mode_is_enabled

Returns whether “Virtualize x2APIC mode” is currently enabled.

virtualize_x2apic_mode_is_enabled_in_value

Returns whether “Virtualize x2APIC mode” is enabled in the given register value.

vmcs_shadowing_is_disabled

Returns whether “VMCS shadowing” is currently disabled.

vmcs_shadowing_is_disabled_in_value

Returns whether “VMCS shadowing” is disabled in the given register value.

vmcs_shadowing_is_enabled

Returns whether “VMCS shadowing” is currently enabled.

vmcs_shadowing_is_enabled_in_value

Returns whether “VMCS shadowing” is enabled in the given register value.

wbinvd_exiting_is_disabled

Returns whether “WBINVD exiting” is currently disabled.

wbinvd_exiting_is_disabled_in_value

Returns whether “WBINVD exiting” is disabled in the given register value.

wbinvd_exiting_is_enabled

Returns whether “WBINVD exiting” is currently enabled.

wbinvd_exiting_is_enabled_in_value

Returns whether “WBINVD exiting” is enabled in the given register value.