Module pal::vmcs::vm_entry_controls

VM-entry controls.

The VM-entry controls constitute a 32-bit vector that governs the basic operation of VM entries. Table 24-13 lists the controls supported. See Chapter 24 for how these controls affect VM entries.

Constants

ENCODING

LONG_NAME

NAME

Functions

conceal_vmx_from_pt_is_disabled

Returns whether “Conceal VMX from PT” is currently disabled.

conceal_vmx_from_pt_is_disabled_in_value

Returns whether “Conceal VMX from PT” is disabled in the given register value.

conceal_vmx_from_pt_is_enabled

Returns whether “Conceal VMX from PT” is currently enabled.

conceal_vmx_from_pt_is_enabled_in_value

Returns whether “Conceal VMX from PT” is enabled in the given register value.

deactivate_dualmonitor_treatment_is_disabled

Returns whether “Deactivate dualmonitor treatment” is currently disabled.

deactivate_dualmonitor_treatment_is_disabled_in_value

Returns whether “Deactivate dualmonitor treatment” is disabled in the given register value.

deactivate_dualmonitor_treatment_is_enabled

Returns whether “Deactivate dualmonitor treatment” is currently enabled.

deactivate_dualmonitor_treatment_is_enabled_in_value

Returns whether “Deactivate dualmonitor treatment” is enabled in the given register value.

disable_conceal_vmx_from_pt

Disables “Conceal VMX from PT”.

disable_conceal_vmx_from_pt_in_value

Disables “Conceal VMX from PT” in the given register value, returning the new value.

disable_deactivate_dualmonitor_treatment

Disables “Deactivate dualmonitor treatment”.

disable_deactivate_dualmonitor_treatment_in_value

Disables “Deactivate dualmonitor treatment” in the given register value, returning the new value.

disable_entry_to_smm

Disables “Entry to SMM”.

disable_entry_to_smm_in_value

Disables “Entry to SMM” in the given register value, returning the new value.

disable_ia_32e_mode_guest

Disables “IA-32e mode guest”.

disable_ia_32e_mode_guest_in_value

Disables “IA-32e mode guest” in the given register value, returning the new value.

disable_load_cet_state

Disables “Load CET state”.

disable_load_cet_state_in_value

Disables “Load CET state” in the given register value, returning the new value.

disable_load_debug_controls

Disables “Load debug controls”.

disable_load_debug_controls_in_value

Disables “Load debug controls” in the given register value, returning the new value.

disable_load_ia32_bndcfgs

Disables “Load IA32_BNDCFGS”.

disable_load_ia32_bndcfgs_in_value

Disables “Load IA32_BNDCFGS” in the given register value, returning the new value.

disable_load_ia32_efer

Disables “Load IA32_EFER”.

disable_load_ia32_efer_in_value

Disables “Load IA32_EFER” in the given register value, returning the new value.

disable_load_ia32_pat

Disables “Load IA32_PAT”.

disable_load_ia32_pat_in_value

Disables “Load IA32_PAT” in the given register value, returning the new value.

disable_load_ia32_perf_global_ctrl

Disables “Load IA32_PERF_GLOBAL_CTRL”.

disable_load_ia32_perf_global_ctrl_in_value

Disables “Load IA32_PERF_GLOBAL_CTRL” in the given register value, returning the new value.

disable_load_ia32_rtit_ctl

Disables “Load IA32_RTIT_CTL”.

disable_load_ia32_rtit_ctl_in_value

Disables “Load IA32_RTIT_CTL” in the given register value, returning the new value.

disable_reserved_12

Disables “Reserved bit 12”.

disable_reserved_12_in_value

Disables “Reserved bit 12” in the given register value, returning the new value.

disable_reserved_19

Disables “Reserved bit 19”.

disable_reserved_19_in_value

Disables “Reserved bit 19” in the given register value, returning the new value.

enable_conceal_vmx_from_pt

Enables “Conceal VMX from PT”.

enable_conceal_vmx_from_pt_in_value

Enables “Conceal VMX from PT” in the given register value, returning the new value.

enable_deactivate_dualmonitor_treatment

Enables “Deactivate dualmonitor treatment”.

enable_deactivate_dualmonitor_treatment_in_value

Enables “Deactivate dualmonitor treatment” in the given register value, returning the new value.

enable_entry_to_smm

Enables “Entry to SMM”.

enable_entry_to_smm_in_value

Enables “Entry to SMM” in the given register value, returning the new value.

enable_ia_32e_mode_guest

Enables “IA-32e mode guest”.

enable_ia_32e_mode_guest_in_value

Enables “IA-32e mode guest” in the given register value, returning the new value.

enable_load_cet_state

Enables “Load CET state”.

enable_load_cet_state_in_value

Enables “Load CET state” in the given register value, returning the new value.

enable_load_debug_controls

Enables “Load debug controls”.

enable_load_debug_controls_in_value

Enables “Load debug controls” in the given register value, returning the new value.

enable_load_ia32_bndcfgs

Enables “Load IA32_BNDCFGS”.

enable_load_ia32_bndcfgs_in_value

Enables “Load IA32_BNDCFGS” in the given register value, returning the new value.

enable_load_ia32_efer

Enables “Load IA32_EFER”.

enable_load_ia32_efer_in_value

Enables “Load IA32_EFER” in the given register value, returning the new value.

enable_load_ia32_pat

Enables “Load IA32_PAT”.

enable_load_ia32_pat_in_value

Enables “Load IA32_PAT” in the given register value, returning the new value.

enable_load_ia32_perf_global_ctrl

Enables “Load IA32_PERF_GLOBAL_CTRL”.

enable_load_ia32_perf_global_ctrl_in_value

Enables “Load IA32_PERF_GLOBAL_CTRL” in the given register value, returning the new value.

enable_load_ia32_rtit_ctl

Enables “Load IA32_RTIT_CTL”.

enable_load_ia32_rtit_ctl_in_value

Enables “Load IA32_RTIT_CTL” in the given register value, returning the new value.

enable_reserved_12

Enables “Reserved bit 12”.

enable_reserved_12_in_value

Enables “Reserved bit 12” in the given register value, returning the new value.

enable_reserved_19

Enables “Reserved bit 19”.

enable_reserved_19_in_value

Enables “Reserved bit 19” in the given register value, returning the new value.

entry_to_smm_is_disabled

Returns whether “Entry to SMM” is currently disabled.

entry_to_smm_is_disabled_in_value

Returns whether “Entry to SMM” is disabled in the given register value.

entry_to_smm_is_enabled

Returns whether “Entry to SMM” is currently enabled.

entry_to_smm_is_enabled_in_value

Returns whether “Entry to SMM” is enabled in the given register value.

get

Returns the current value of the register.

get_reserved_1_0

Returns the current value of “Reserved bits 1:0”.

get_reserved_1_0_from_value

Returns the value of “Reserved bits 1:0” in the given register value.

get_reserved_8_3

Returns the current value of “Reserved bits 8:3”.

get_reserved_8_3_from_value

Returns the value of “Reserved bits 8:3” in the given register value.

get_reserved_31_21

Returns the current value of “Reserved bits 31:21”.

get_reserved_31_21_from_value

Returns the value of “Reserved bits 31:21” in the given register value.

ia_32e_mode_guest_is_disabled

Returns whether “IA-32e mode guest” is currently disabled.

ia_32e_mode_guest_is_disabled_in_value

Returns whether “IA-32e mode guest” is disabled in the given register value.

ia_32e_mode_guest_is_enabled

Returns whether “IA-32e mode guest” is currently enabled.

ia_32e_mode_guest_is_enabled_in_value

Returns whether “IA-32e mode guest” is enabled in the given register value.

load_cet_state_is_disabled

Returns whether “Load CET state” is currently disabled.

load_cet_state_is_disabled_in_value

Returns whether “Load CET state” is disabled in the given register value.

load_cet_state_is_enabled

Returns whether “Load CET state” is currently enabled.

load_cet_state_is_enabled_in_value

Returns whether “Load CET state” is enabled in the given register value.

load_debug_controls_is_disabled

Returns whether “Load debug controls” is currently disabled.

load_debug_controls_is_disabled_in_value

Returns whether “Load debug controls” is disabled in the given register value.

load_debug_controls_is_enabled

Returns whether “Load debug controls” is currently enabled.

load_debug_controls_is_enabled_in_value

Returns whether “Load debug controls” is enabled in the given register value.

load_ia32_bndcfgs_is_disabled

Returns whether “Load IA32_BNDCFGS” is currently disabled.

load_ia32_bndcfgs_is_disabled_in_value

Returns whether “Load IA32_BNDCFGS” is disabled in the given register value.

load_ia32_bndcfgs_is_enabled

Returns whether “Load IA32_BNDCFGS” is currently enabled.

load_ia32_bndcfgs_is_enabled_in_value

Returns whether “Load IA32_BNDCFGS” is enabled in the given register value.

load_ia32_efer_is_disabled

Returns whether “Load IA32_EFER” is currently disabled.

load_ia32_efer_is_disabled_in_value

Returns whether “Load IA32_EFER” is disabled in the given register value.

load_ia32_efer_is_enabled

Returns whether “Load IA32_EFER” is currently enabled.

load_ia32_efer_is_enabled_in_value

Returns whether “Load IA32_EFER” is enabled in the given register value.

load_ia32_pat_is_disabled

Returns whether “Load IA32_PAT” is currently disabled.

load_ia32_pat_is_disabled_in_value

Returns whether “Load IA32_PAT” is disabled in the given register value.

load_ia32_pat_is_enabled

Returns whether “Load IA32_PAT” is currently enabled.

load_ia32_pat_is_enabled_in_value

Returns whether “Load IA32_PAT” is enabled in the given register value.

load_ia32_perf_global_ctrl_is_disabled

Returns whether “Load IA32_PERF_GLOBAL_CTRL” is currently disabled.

load_ia32_perf_global_ctrl_is_disabled_in_value

Returns whether “Load IA32_PERF_GLOBAL_CTRL” is disabled in the given register value.

load_ia32_perf_global_ctrl_is_enabled

Returns whether “Load IA32_PERF_GLOBAL_CTRL” is currently enabled.

load_ia32_perf_global_ctrl_is_enabled_in_value

Returns whether “Load IA32_PERF_GLOBAL_CTRL” is enabled in the given register value.

load_ia32_rtit_ctl_is_disabled

Returns whether “Load IA32_RTIT_CTL” is currently disabled.

load_ia32_rtit_ctl_is_disabled_in_value

Returns whether “Load IA32_RTIT_CTL” is disabled in the given register value.

load_ia32_rtit_ctl_is_enabled

Returns whether “Load IA32_RTIT_CTL” is currently enabled.

load_ia32_rtit_ctl_is_enabled_in_value

Returns whether “Load IA32_RTIT_CTL” is enabled in the given register value.

reserved_12_is_disabled

Returns whether “Reserved bit 12” is currently disabled.

reserved_12_is_disabled_in_value

Returns whether “Reserved bit 12” is disabled in the given register value.

reserved_12_is_enabled

Returns whether “Reserved bit 12” is currently enabled.

reserved_12_is_enabled_in_value

Returns whether “Reserved bit 12” is enabled in the given register value.

reserved_19_is_disabled

Returns whether “Reserved bit 19” is currently disabled.

reserved_19_is_disabled_in_value

Returns whether “Reserved bit 19” is disabled in the given register value.

reserved_19_is_enabled

Returns whether “Reserved bit 19” is currently enabled.

reserved_19_is_enabled_in_value

Returns whether “Reserved bit 19” is enabled in the given register value.

set

Sets the value of the register.

set_reserved_1_0

Sets the value of “Reserved bits 1:0”.

set_reserved_1_0_in_value

Sets the value of “Reserved bits 1:0” in the given register value, returning the new value.

set_reserved_8_3

Sets the value of “Reserved bits 8:3”.

set_reserved_8_3_in_value

Sets the value of “Reserved bits 8:3” in the given register value, returning the new value.

set_reserved_31_21

Sets the value of “Reserved bits 31:21”.

set_reserved_31_21_in_value

Sets the value of “Reserved bits 31:21” in the given register value, returning the new value.