Function pal::vmcs::vm_entry_controls::set_reserved_1_0_in_value

pub fn set_reserved_1_0_in_value(field_value: u32, register_value: &mut u32)

Sets the value of “Reserved bits 1:0” in the given register value, returning the new value.

All other bits in this field are reserved, some to 0 and some to 1. Software should consult the VMX capability MSRs IA32_VMX_ENTRY_CTLS and IA32_VMX_TRUE_ENTRY_CTLS (see Appendix A.5) to determine how it should set the reserved bits. Failure to set reserved bits properly causes subsequent VM entries to fail (see Section 26.2.1.3).