Module pal::vmcs::vm_exit_controls

VM-exit controls.

The VM-exit controls constitute a 32-bit vector that governs the basic operation of VM exits. Table 24-11 lists the controls supported. See Chapter 27 for complete details of how these controls affect VM exits

Constants

ENCODING

LONG_NAME

NAME

Functions

acknowledge_interrupt_on_exit_is_disabled

Returns whether “Acknowledge interrupt on exit” is currently disabled.

acknowledge_interrupt_on_exit_is_disabled_in_value

Returns whether “Acknowledge interrupt on exit” is disabled in the given register value.

acknowledge_interrupt_on_exit_is_enabled

Returns whether “Acknowledge interrupt on exit” is currently enabled.

acknowledge_interrupt_on_exit_is_enabled_in_value

Returns whether “Acknowledge interrupt on exit” is enabled in the given register value.

clear_ia32_bndcfgs_is_disabled

Returns whether “Clear IA32_BNDCFGS” is currently disabled.

clear_ia32_bndcfgs_is_disabled_in_value

Returns whether “Clear IA32_BNDCFGS” is disabled in the given register value.

clear_ia32_bndcfgs_is_enabled

Returns whether “Clear IA32_BNDCFGS” is currently enabled.

clear_ia32_bndcfgs_is_enabled_in_value

Returns whether “Clear IA32_BNDCFGS” is enabled in the given register value.

clear_ia32_rtit_ctl_is_disabled

Returns whether “Clear IA32_RTIT_CTL” is currently disabled.

clear_ia32_rtit_ctl_is_disabled_in_value

Returns whether “Clear IA32_RTIT_CTL” is disabled in the given register value.

clear_ia32_rtit_ctl_is_enabled

Returns whether “Clear IA32_RTIT_CTL” is currently enabled.

clear_ia32_rtit_ctl_is_enabled_in_value

Returns whether “Clear IA32_RTIT_CTL” is enabled in the given register value.

conceal_vmx_from_pt_is_disabled

Returns whether “Conceal VMX from PT” is currently disabled.

conceal_vmx_from_pt_is_disabled_in_value

Returns whether “Conceal VMX from PT” is disabled in the given register value.

conceal_vmx_from_pt_is_enabled

Returns whether “Conceal VMX from PT” is currently enabled.

conceal_vmx_from_pt_is_enabled_in_value

Returns whether “Conceal VMX from PT” is enabled in the given register value.

disable_acknowledge_interrupt_on_exit

Disables “Acknowledge interrupt on exit”.

disable_acknowledge_interrupt_on_exit_in_value

Disables “Acknowledge interrupt on exit” in the given register value, returning the new value.

disable_clear_ia32_bndcfgs

Disables “Clear IA32_BNDCFGS”.

disable_clear_ia32_bndcfgs_in_value

Disables “Clear IA32_BNDCFGS” in the given register value, returning the new value.

disable_clear_ia32_rtit_ctl

Disables “Clear IA32_RTIT_CTL”.

disable_clear_ia32_rtit_ctl_in_value

Disables “Clear IA32_RTIT_CTL” in the given register value, returning the new value.

disable_conceal_vmx_from_pt

Disables “Conceal VMX from PT”.

disable_conceal_vmx_from_pt_in_value

Disables “Conceal VMX from PT” in the given register value, returning the new value.

disable_host_address_space_size

Disables “Host address-space size”.

disable_host_address_space_size_in_value

Disables “Host address-space size” in the given register value, returning the new value.

disable_load_cet_state

Disables “Load CET state”.

disable_load_cet_state_in_value

Disables “Load CET state” in the given register value, returning the new value.

disable_load_ia32_efer

Disables “Load IA32_EFER”.

disable_load_ia32_efer_in_value

Disables “Load IA32_EFER” in the given register value, returning the new value.

disable_load_ia32_pat

Disables “Load IA32_PAT”.

disable_load_ia32_pat_in_value

Disables “Load IA32_PAT” in the given register value, returning the new value.

disable_load_ia32_perf_global_ctrl

Disables “Load IA32_PERF_GLOBAL_CTRL”.

disable_load_ia32_perf_global_ctrl_in_value

Disables “Load IA32_PERF_GLOBAL_CTRL” in the given register value, returning the new value.

disable_save_debug_controls

Disables “Save debug controls”.

disable_save_debug_controls_in_value

Disables “Save debug controls” in the given register value, returning the new value.

disable_save_ia32_efer

Disables “Save IA32_EFER”.

disable_save_ia32_efer_in_value

Disables “Save IA32_EFER” in the given register value, returning the new value.

disable_save_ia32_pat

Disables “Save IA32_PAT”.

disable_save_ia32_pat_in_value

Disables “Save IA32_PAT” in the given register value, returning the new value.

disable_save_vmxpreemption_timer_value

Disables “Save VMXpreemption timer value”.

disable_save_vmxpreemption_timer_value_in_value

Disables “Save VMXpreemption timer value” in the given register value, returning the new value.

enable_acknowledge_interrupt_on_exit

Enables “Acknowledge interrupt on exit”.

enable_acknowledge_interrupt_on_exit_in_value

Enables “Acknowledge interrupt on exit” in the given register value, returning the new value.

enable_clear_ia32_bndcfgs

Enables “Clear IA32_BNDCFGS”.

enable_clear_ia32_bndcfgs_in_value

Enables “Clear IA32_BNDCFGS” in the given register value, returning the new value.

enable_clear_ia32_rtit_ctl

Enables “Clear IA32_RTIT_CTL”.

enable_clear_ia32_rtit_ctl_in_value

Enables “Clear IA32_RTIT_CTL” in the given register value, returning the new value.

enable_conceal_vmx_from_pt

Enables “Conceal VMX from PT”.

enable_conceal_vmx_from_pt_in_value

Enables “Conceal VMX from PT” in the given register value, returning the new value.

enable_host_address_space_size

Enables “Host address-space size”.

enable_host_address_space_size_in_value

Enables “Host address-space size” in the given register value, returning the new value.

enable_load_cet_state

Enables “Load CET state”.

enable_load_cet_state_in_value

Enables “Load CET state” in the given register value, returning the new value.

enable_load_ia32_efer

Enables “Load IA32_EFER”.

enable_load_ia32_efer_in_value

Enables “Load IA32_EFER” in the given register value, returning the new value.

enable_load_ia32_pat

Enables “Load IA32_PAT”.

enable_load_ia32_pat_in_value

Enables “Load IA32_PAT” in the given register value, returning the new value.

enable_load_ia32_perf_global_ctrl

Enables “Load IA32_PERF_GLOBAL_CTRL”.

enable_load_ia32_perf_global_ctrl_in_value

Enables “Load IA32_PERF_GLOBAL_CTRL” in the given register value, returning the new value.

enable_save_debug_controls

Enables “Save debug controls”.

enable_save_debug_controls_in_value

Enables “Save debug controls” in the given register value, returning the new value.

enable_save_ia32_efer

Enables “Save IA32_EFER”.

enable_save_ia32_efer_in_value

Enables “Save IA32_EFER” in the given register value, returning the new value.

enable_save_ia32_pat

Enables “Save IA32_PAT”.

enable_save_ia32_pat_in_value

Enables “Save IA32_PAT” in the given register value, returning the new value.

enable_save_vmxpreemption_timer_value

Enables “Save VMXpreemption timer value”.

enable_save_vmxpreemption_timer_value_in_value

Enables “Save VMXpreemption timer value” in the given register value, returning the new value.

get

Returns the current value of the register.

get_reserved_1_0

Returns the current value of “Reserved bits 1:0”.

get_reserved_1_0_from_value

Returns the value of “Reserved bits 1:0” in the given register value.

get_reserved_8_3

Returns the current value of “Reserved bits 8:3”.

get_reserved_8_3_from_value

Returns the value of “Reserved bits 8:3” in the given register value.

get_reserved_12_10

Returns the current value of “Reserved bits 12:10”.

get_reserved_12_10_from_value

Returns the value of “Reserved bits 12:10” in the given register value.

get_reserved_14_13

Returns the current value of “Reserved bits 14:13”.

get_reserved_14_13_from_value

Returns the value of “Reserved bits 14:13” in the given register value.

get_reserved_17_16

Returns the current value of “Reserved bits 17:16”.

get_reserved_17_16_from_value

Returns the value of “Reserved bits 17:16” in the given register value.

get_reserved_27_26

Returns the current value of “Reserved bits 27:26”.

get_reserved_27_26_from_value

Returns the value of “Reserved bits 27:26” in the given register value.

get_reserved_31_29

Returns the current value of “Reserved bits 31:29”.

get_reserved_31_29_from_value

Returns the value of “Reserved bits 31:29” in the given register value.

host_address_space_size_is_disabled

Returns whether “Host address-space size” is currently disabled.

host_address_space_size_is_disabled_in_value

Returns whether “Host address-space size” is disabled in the given register value.

host_address_space_size_is_enabled

Returns whether “Host address-space size” is currently enabled.

host_address_space_size_is_enabled_in_value

Returns whether “Host address-space size” is enabled in the given register value.

load_cet_state_is_disabled

Returns whether “Load CET state” is currently disabled.

load_cet_state_is_disabled_in_value

Returns whether “Load CET state” is disabled in the given register value.

load_cet_state_is_enabled

Returns whether “Load CET state” is currently enabled.

load_cet_state_is_enabled_in_value

Returns whether “Load CET state” is enabled in the given register value.

load_ia32_efer_is_disabled

Returns whether “Load IA32_EFER” is currently disabled.

load_ia32_efer_is_disabled_in_value

Returns whether “Load IA32_EFER” is disabled in the given register value.

load_ia32_efer_is_enabled

Returns whether “Load IA32_EFER” is currently enabled.

load_ia32_efer_is_enabled_in_value

Returns whether “Load IA32_EFER” is enabled in the given register value.

load_ia32_pat_is_disabled

Returns whether “Load IA32_PAT” is currently disabled.

load_ia32_pat_is_disabled_in_value

Returns whether “Load IA32_PAT” is disabled in the given register value.

load_ia32_pat_is_enabled

Returns whether “Load IA32_PAT” is currently enabled.

load_ia32_pat_is_enabled_in_value

Returns whether “Load IA32_PAT” is enabled in the given register value.

load_ia32_perf_global_ctrl_is_disabled

Returns whether “Load IA32_PERF_GLOBAL_CTRL” is currently disabled.

load_ia32_perf_global_ctrl_is_disabled_in_value

Returns whether “Load IA32_PERF_GLOBAL_CTRL” is disabled in the given register value.

load_ia32_perf_global_ctrl_is_enabled

Returns whether “Load IA32_PERF_GLOBAL_CTRL” is currently enabled.

load_ia32_perf_global_ctrl_is_enabled_in_value

Returns whether “Load IA32_PERF_GLOBAL_CTRL” is enabled in the given register value.

save_debug_controls_is_disabled

Returns whether “Save debug controls” is currently disabled.

save_debug_controls_is_disabled_in_value

Returns whether “Save debug controls” is disabled in the given register value.

save_debug_controls_is_enabled

Returns whether “Save debug controls” is currently enabled.

save_debug_controls_is_enabled_in_value

Returns whether “Save debug controls” is enabled in the given register value.

save_ia32_efer_is_disabled

Returns whether “Save IA32_EFER” is currently disabled.

save_ia32_efer_is_disabled_in_value

Returns whether “Save IA32_EFER” is disabled in the given register value.

save_ia32_efer_is_enabled

Returns whether “Save IA32_EFER” is currently enabled.

save_ia32_efer_is_enabled_in_value

Returns whether “Save IA32_EFER” is enabled in the given register value.

save_ia32_pat_is_disabled

Returns whether “Save IA32_PAT” is currently disabled.

save_ia32_pat_is_disabled_in_value

Returns whether “Save IA32_PAT” is disabled in the given register value.

save_ia32_pat_is_enabled

Returns whether “Save IA32_PAT” is currently enabled.

save_ia32_pat_is_enabled_in_value

Returns whether “Save IA32_PAT” is enabled in the given register value.

save_vmxpreemption_timer_value_is_disabled

Returns whether “Save VMXpreemption timer value” is currently disabled.

save_vmxpreemption_timer_value_is_disabled_in_value

Returns whether “Save VMXpreemption timer value” is disabled in the given register value.

save_vmxpreemption_timer_value_is_enabled

Returns whether “Save VMXpreemption timer value” is currently enabled.

save_vmxpreemption_timer_value_is_enabled_in_value

Returns whether “Save VMXpreemption timer value” is enabled in the given register value.

set

Sets the value of the register.

set_reserved_1_0

Sets the value of “Reserved bits 1:0”.

set_reserved_1_0_in_value

Sets the value of “Reserved bits 1:0” in the given register value, returning the new value.

set_reserved_8_3

Sets the value of “Reserved bits 8:3”.

set_reserved_8_3_in_value

Sets the value of “Reserved bits 8:3” in the given register value, returning the new value.

set_reserved_12_10

Sets the value of “Reserved bits 12:10”.

set_reserved_12_10_in_value

Sets the value of “Reserved bits 12:10” in the given register value, returning the new value.

set_reserved_14_13

Sets the value of “Reserved bits 14:13”.

set_reserved_14_13_in_value

Sets the value of “Reserved bits 14:13” in the given register value, returning the new value.

set_reserved_17_16

Sets the value of “Reserved bits 17:16”.

set_reserved_17_16_in_value

Sets the value of “Reserved bits 17:16” in the given register value, returning the new value.

set_reserved_27_26

Sets the value of “Reserved bits 27:26”.

set_reserved_27_26_in_value

Sets the value of “Reserved bits 27:26” in the given register value, returning the new value.

set_reserved_31_29

Sets the value of “Reserved bits 31:29”.

set_reserved_31_29_in_value

Sets the value of “Reserved bits 31:29” in the given register value, returning the new value.