Module pal::vmcs::vm_function_controls

VM-function controls (full).

The VM-function controls constitute a 64-bit vector that governs use of the VMFUNC instruction in VMX non-root operation. This field is supported only on processors that support the 1-settings of both the “activate secondary controls” primary processor-based VM-execution control and the “enable VM functions” secondary processorbased VM- execution control. Table 24-9 lists the VM-function controls. See Section 25.5.5 for more details of how these controls affect processor behavior in VMX non-root operation.

Constants

ENCODING

LONG_NAME

NAME

Functions

disable_eptp_switching

Disables “EPTP switching”.

disable_eptp_switching_in_value

Disables “EPTP switching” in the given register value, returning the new value.

enable_eptp_switching

Enables “EPTP switching”.

enable_eptp_switching_in_value

Enables “EPTP switching” in the given register value, returning the new value.

eptp_switching_is_disabled

Returns whether “EPTP switching” is currently disabled.

eptp_switching_is_disabled_in_value

Returns whether “EPTP switching” is disabled in the given register value.

eptp_switching_is_enabled

Returns whether “EPTP switching” is currently enabled.

eptp_switching_is_enabled_in_value

Returns whether “EPTP switching” is enabled in the given register value.

get

Returns the current value of the register.

set

Sets the value of the register.